% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add}
\alias{add}
\alias{add.Container}
\alias{add.Dict}
\alias{add.dict.table}
\alias{addleft}
\alias{addleft.Deque}
\title{Add elements}
\usage{
add(x, ...)

\method{add}{Container}(x, ...)

\method{add}{Dict}(x, ...)

\method{add}{dict.table}(x, ...)

addleft(x, ...)

\method{addleft}{Deque}(x, ...)
}
\arguments{
\item{x}{an \code{R} object of the respective class.}

\item{...}{elements to be added. If \code{x} is a \code{Dict} or \code{dict.table} object,
all elements must be named, that is, of the form \code{key = value}.}
}
\value{
For \code{Container}, an object of class \code{Container} (or of the
respective derived class) with the elements being added to \code{x}.

For \code{Dict}, an object of class \code{Dict} with the key-value pairs being
added to \code{x}. If one of the keys already exists, an error is given.

For \code{dict.table}, an object of class \code{dict.table} with the columns
being added to \code{x}. All given columns must be named. If one of the column
names already exists, an error is given.

For \code{Deque}, an object of class \code{Deque} with the elements being
added to the right or left (\code{addleft}) of \code{x}.
}
\description{
Add elements to container-like objects.
}
\examples{
co = container(1)
add(co, 1, b = 2, c = container(1:3))

s = setnew(1)
add(s, 1, 1, b = 2, "1", co = container(1, 1))

d = dict(a = 1)
add(d, b = 2, co = container(1:3))

\dontrun{
add(d, a = 7:9)  # key 'a' already in Dict}

dit = dict.table(a = 1:3)
add(dit, b = 3:1, d = 4:6)

\dontrun{
add(dit, a = 7:9)  # column 'a' already exists}

d = deque(0)
add(d, a = 1, b = 2)         # |0, a = 1, b = 2|
addleft(d, a = 1, b = 2)     # |b = 2, a = 1, 0|
}
