% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear.R
\name{clear}
\alias{clear}
\alias{clear.Container}
\alias{clear.dict.table}
\title{Clear a container}
\usage{
clear(x, ...)

\method{clear}{Container}(x)

\method{clear}{dict.table}(x)
}
\arguments{
\item{x}{any \code{R} object.}

\item{...}{additional arguments to be passed to or from methods.}
}
\value{
For \code{Container}, an object of class \code{Container} with all elements
being removed.

For \code{dict.table}, an object of class \code{dict.table} with all elements
being removed.
}
\description{
Removes all elements from the container object.
}
