% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clone.R
\name{clone}
\alias{clone}
\title{Clone an object}
\usage{
clone(x, ...)
}
\arguments{
\item{x}{any \code{R} object.}

\item{...}{additional arguments to be passed to or from methods.}
}
\value{
A copy of the object.
}
\description{
Creates a copy of the object.
}
