% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container-deprecated.R
\name{container-deprecated}
\alias{container-deprecated}
\alias{empty}
\alias{empty.Container}
\alias{size}
\alias{size.Container}
\alias{sortkey}
\alias{sortkey.Dict}
\alias{values}
\alias{values.Container}
\alias{values.dict.table}
\alias{keys}
\title{Deprecated Functions in Package container}
\usage{
empty(x)

\method{empty}{Container}(x)

size(x)

\method{size}{Container}(x)

sortkey(x, decr = FALSE)

\method{sortkey}{Dict}(x, decr = FALSE)

values(x)

\method{values}{Container}(x)

\method{values}{dict.table}(x)

keys(x)
}
\arguments{
\item{x}{any \code{R} object.}

\item{decr}{\code{logical} sort decreasingly?}

\item{...}{(arguments)}
}
\description{
These functions are provided for backwards-compatibility and
may be defunct as soon as the next release.
}
\details{
empty(...)       ### -- use \code{is_empty(...)} instead
set(...)         ### -- use \code{setnew(...)} instead to create \code{Set}s
size(...)        ### -- use \code{length(...)} instead
sortkey(...)     ### -- \code{Dict} keys are now always sorted
remove(...)      ### -- use \code{delete(...)} instead
type(...)        ### -- not of use anymore
values(...)      ### -- use \code{as.list(...)} instead
}
