% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{container_options}
\alias{container_options}
\title{Set container package options}
\usage{
container_options(..., .reset = FALSE)
}
\arguments{
\item{...}{any options can be defined, using name = value.}

\item{.reset}{\code{logical} if TRUE, the options are reset to their default and
returned.}
}
\value{
\itemize{
\item container_options() returns a list of all set options sorted by name.
\item container_options(name), a list of length one containing the set value,
or NULL if it is unset. Can also be multiple names (see examples).
\item container_options(key = value) sets the option with name \code{key} to \code{value}
and returns the previous options invisibly.
}
}
\description{
Set container package options
}
\examples{
container_options()
old <- container_options(dots = FALSE)
old
container_options("dots")
container_options("dots", "vec.len")
container_options(cmp = "identical", foo = "bar")
container_options()
(container_options(.reset = TRUE))
}
