% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container-S3.R
\name{ContainerS3op}
\alias{ContainerS3op}
\alias{+.Container}
\title{Container operators}
\usage{
\method{+}{Container}(c1, c2)
}
\arguments{
\item{c1}{\code{\link[container]{Container}} object}

\item{c2}{\code{\link[container]{Container}} object}
}
\value{
\code{\link[container]{Container}} object
}
\description{
Binary operators for \code{Container} objects.
}
\details{
\code{c1 + c2}: return \code{c1} and \code{c2} combined (as a copy)
}
