% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{count}
\alias{count}
\alias{count.Deque}
\alias{count.Set}
\title{Count elements}
\usage{
count(x, elem, ...)

\method{count}{Deque}(x, elem)

\method{count}{Set}(x, elem)
}
\arguments{
\item{x}{any \code{R} object.}

\item{elem}{element to counted.}

\item{...}{additional arguments to be passed to or from methods.}
}
\value{
Returns how many times \code{elem} occurs in the object.
}
\description{
Count the number of occurences of some element.
}
