% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.R
\name{delete}
\alias{delete}
\alias{delete_}
\alias{delete.Container}
\alias{delete_.Container}
\title{Delete elements}
\usage{
delete(.x, ...)

delete_(.x, ...)

\method{delete}{Container}(.x, ...)

\method{delete_}{Container}(.x, ...)
}
\arguments{
\item{.x}{any \code{R} object.}

\item{...}{elements to be deleted.}
}
\value{
For \code{Container}, an object of class \code{Container} (or one of the
respective derived classes).
}
\description{
Search and remove elements from an object. If the element is not found,
an error is signaled.
}
\examples{

s = setnew("a", 1:3, iris)
print(s)
delete(s, 1:3, "a")
delete(s, iris)
\dontrun{
delete(s, "b")  # "b" is not in Set}
}
