% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container-deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{empty}
\alias{empty.Container}
\alias{size}
\alias{size.Container}
\alias{sortkey}
\alias{sortkey.Dict}
\alias{values}
\alias{values.Container}
\alias{values.dict.table}
\alias{keys}
\alias{set}
\title{Deprecated Functions}
\usage{
empty(x)

\method{empty}{Container}(x)

size(x)

\method{size}{Container}(x)

sortkey(x, decr = FALSE)

\method{sortkey}{Dict}(x, decr = FALSE)

values(x)

\method{values}{Container}(x)

\method{values}{dict.table}(x)

keys(x)

set(...)
}
\arguments{
\item{x}{any \code{R} object.}

\item{decr}{\code{logical} sort decreasingly?}

\item{...}{additional arguments.}
}
\description{
These functions are provided for backwards-compatibility and
may be defunct as soon as the next release.
}
\details{
\itemize{
\item \code{\link[=empty]{empty()}}      \code{\link[=is_empty]{is_empty()}} instead
\item \code{\link[=set]{set()}}        \code{\link[=setnew]{setnew()}} instead
\item \code{\link[=size]{size()}}       use \code{\link[=length]{length()}} instead
\item \code{\link[=sortkey]{sortkey()}}    keys of \link{Dict} objects are now always sorted
\item \code{\link[=remove]{remove()}}     use \code{\link[=delete]{delete()}} instead
\item \code{\link[=type]{type()}}       not of use anymore
\item \code{\link[=values]{values()}}     use \code{\link[=as.list]{as.list()}} instead
}
}
