% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deque-S3.R
\name{DequeDictS3funcs}
\alias{DequeDictS3funcs}
\alias{peek}
\alias{pop}
\title{Deque and Dict S3 member functions}
\usage{
peek(x, ...)

pop(x, ...)
}
\arguments{
\item{x}{object of class \code{Deque} or \code{Dict}}

\item{...}{further arguments}
}
\description{
Access elements from \code{Deque} or \code{Dict} objects.
}
