% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deque-S3.R
\name{dequeS3binOp}
\alias{dequeS3binOp}
\alias{+.Deque}
\title{Binary deque operators}
\usage{
\method{+}{Deque}(x1, x2)
}
\arguments{
\item{x1}{primitive or \code{\link[container]{Deque}} object}

\item{x2}{primitive or \code{\link[container]{Deque}} object}
}
\value{
\code{\link[container]{Deque}} object
}
\description{
Binary operators for \code{Deque} objects.
}
\details{
\code{x1 + x2}:
}
