% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict.frame-S3.R
\name{dict.frameS3}
\alias{dict.frameS3}
\alias{dict.frame}
\alias{as.dict.frame}
\alias{is.dict.frame}
\title{Dict S3 interface}
\usage{
dict.frame(...)

as.dict.frame(x)

is.dict.frame(x)
}
\arguments{
\item{x}{a named vector (or list) of 'any' type including
\code{\link[base:data.frame]{base::data.frame()}}s.}
}
\description{
The \code{\link[=Dict.frame]{Dict.frame()}} is a mix of a dictionary and a
\code{data.frame}. In particular, it is a dictionary thereby inheriting all
\code{\link[=Dict]{Dict()}} methods, but where each element has the same length. In contrast to
a \code{\link[base:data.frame]{base::data.frame()}}, a \code{\link[=Dict.frame]{Dict.frame()}} can contain arbitrary complex
objects.
}
\details{
For a detailed documentation of all methods see \code{\link[=Dict.frame]{Dict.frame()}}.
}
\seealso{
\code{\link[=Dict]{Dict()}}, \code{\link[base:data.frame]{base::data.frame()}}
}
