% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict.table-S3.R
\name{dict.table.replace}
\alias{dict.table.replace}
\title{Extract or replace parts of a \code{dict.table}}
\arguments{
\item{x}{\code{\link[=dict.table]{dict.table()}} object.}

\item{i}{\code{integer} row indices of elements to extract or replace.}

\item{j}{\code{integer} or \code{character} column indices of elements to extract
or replace.}
}
\description{
Access and assignment operators for \code{\link[=dict.table]{dict.table()}} objects.
}
