% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container-deprecated.R
\name{empty-deprecated}
\alias{empty-deprecated}
\alias{empty}
\alias{empty.Container}
\title{Check if object is empty}
\usage{
empty(x)

\method{empty}{Container}(x)
}
\arguments{
\item{x}{any \code{R} object.}
}
\value{
\code{TRUE} if object is empty otherwise \code{FALSE}.
}
\description{
Check if object is empty
}
