% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empty.R
\name{empty}
\alias{empty}
\alias{empty.Container}
\alias{empty.dict.table}
\title{Check if object is empty}
\usage{
empty(x, ...)

\method{empty}{Container}(x)

\method{empty}{dict.table}(x)
}
\arguments{
\item{x}{any \code{R} object.}

\item{...}{additional arguments to be passed to or from methods.}
}
\value{
\code{TRUE} if object is empty otherwise \code{FALSE}.
}
\description{
Check if object is empty
}
