% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ops-extract.R
\name{ExtractContainer}
\alias{ExtractContainer}
\alias{[.Container}
\alias{[<-.Container}
\title{Extract or Replace Parts of a Container Object}
\usage{
x[i]
x[[i]]
x[...]

x[i] <- value
x[[i]] <- value
x$name <- value
}
\arguments{
\item{x}{\code{Container} object for which to extract or replace elements.}

\item{name}{\code{character} string (possibly backtick quoted)}

\item{value}{the replacing value of \code{ANY} type}

\item{i, ...}{indices specifying elements to extract or replace. Indices
are \code{numeric} or \code{character} vectors or a \code{list} containing both.}
}
\description{
Extract or replace parts of a \code{Container} object similar
to R's base extract operators on lists.
}
