% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_at.R
\name{get_at}
\alias{get_at}
\alias{get_at.Dict}
\alias{get_at.dict.table}
\title{Strict element access}
\usage{
get_at(.x, ...)

\method{get_at}{Dict}(.x, ...)

\method{get_at}{dict.table}(.x, ...)
}
\arguments{
\item{.x}{any \code{R} object.}

\item{...}{names or indices to be accessed and additional arguments to be
passed to or from methods.}
}
\value{
For \code{Dict}, a \code{Dict} object containing the selected elements.

For \code{dict.table}, a \code{dict.table} object containing the selected
columns.
}
\description{
Try to access elements and signal an error if not found.
}
\examples{

d = dict(a = 1, b = 1:3)
get_at(d, "b")
get_at(d, "a", "b")      # or alternatively:
get_at(d, c("a", "b"))
\dontrun{
get_at(d, "z")   # key 'z' not in Dict}

dit = dict.table(a = 1:3, b = 4:6)
get_at(dit, "b")
get_at(dit, "a", "b")    # or alternatively:
get_at(dit, c("a", "b"))
}
\seealso{
\code{\link[=peek]{peek()}} for less strict element access.
}
