% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_label.R
\name{get_label}
\alias{get_label}
\title{Object labels}
\usage{
get_label(x, vec.len = 4L, ...)
}
\arguments{
\item{x}{\code{ANY} R object}

\item{vec.len}{length of vector that is still printed}

\item{...}{arguments passed to and from methods}
}
\description{
Object labels
}
\references{
Credits for the code idea below belong to the authors of the
very well written \code{sets} package.
}
\seealso{
\code{LABEL} function of the \code{sets} package.
}
