% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getvalues.R
\name{getvalues}
\alias{getvalues}
\alias{getvalues.Dict}
\alias{getvalues.dict.table}
\title{Strict multiple element access}
\usage{
getvalues(x, ...)

\method{getvalues}{Dict}(x, ...)

\method{getvalues}{dict.table}(x, ...)
}
\arguments{
\item{x}{any \code{R} object.}

\item{...}{names or indices associated with the desired elements.}
}
\value{
A new \code{Dict} containing the desired the values.
}
\description{
Try to access elements and signal an error if any is not found.
This function is provided as a convenience function for interactive use to
enable easy access of multiple elements.
}
\examples{


}
\seealso{
\code{\link[=getval]{getval()}}
}
