% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-IteratorR6.R
\name{is.iterable}
\alias{is.iterable}
\title{Check if object is iterable}
\usage{
is.iterable(x)
}
\arguments{
\item{any}{\code{R} object}
}
\value{
returns \code{TRUE} if object is iterable otherwise \code{FALSE}
}
\description{
Check if object is iterable
}
