% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_empty.R
\name{is_empty}
\alias{is_empty}
\alias{is_empty.Container}
\alias{is_empty.dict.table}
\title{Check if object is empty}
\usage{
is_empty(x, ...)

\method{is_empty}{Container}(x)

\method{is_empty}{dict.table}(x)
}
\arguments{
\item{x}{any \code{R} object.}

\item{...}{additional arguments to be passed to or from methods.}
}
\value{
\code{TRUE} if object is empty otherwise \code{FALSE}.
}
\description{
Check if object is empty
}
\examples{

co = container(1, 2)
is_empty(co)
is_empty(co$clear())

d = dict.table(a = 1:4, b = 4:1)
is_empty(d)
is_empty(d$clear())
}
