% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ops-relational.R
\name{OpsComp}
\alias{OpsComp}
\title{Relational Operators for Containers}
\arguments{
\item{x, y}{objects of class \code{\link[=Container]{Container()}} or one of the derived classes.}
}
\description{
Binary comparison operators for \code{\link[=Container]{Container()}} objects and
derived classes.
}
