% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ops-logic.R
\name{OpsLogic}
\alias{OpsLogic}
\alias{&.Dict}
\alias{|.Dict}
\alias{&.Set}
\alias{|.Set}
\title{Logic Operators for Containers}
\usage{
\method{&}{Dict}(x, y)

\method{|}{Dict}(x, y)

\method{&}{Set}(x, y)

\method{|}{Set}(x, y)
}
\arguments{
\item{x, y}{Depending on the operator at least one must be of class
\code{\link[=Container]{Container()}} or the respective derived class and the other at least be
coercible to the respective class.}
}
\value{
For \code{Dict}, \code{&} returns a copy of \code{x} keeping only the keys that
are common in both (key intersection), that is, all keys in \code{x} that do not
exist in \code{y} were removed.

For \code{Dict}, \code{|} returns a copy of \code{x} extended by all elements of
\code{y} that are stored at keys (or names) that do not exist in \code{x}, thereby
combining the keys of both objects (i.e. basically union of keys).

For \code{Set}, \code{&} returns the set intersection of x and y

For \code{Set}, \code{|} returns the set union of x and y
}
\description{
Binary logic operators for \code{\link[=Container]{Container()}} objects and
derived classes.
}
