% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peek.R
\name{peekitem}
\alias{peekitem}
\alias{peekitem.Container}
\title{Peek random element}
\usage{
peekitem(x)

\method{peekitem}{Container}(x)
}
\arguments{
\item{x}{any \code{R} object.}

\item{...}{additional arguments to be passed to or from methods.}
}
\value{
For \code{Container} a randomly peeked element is returned.
}
\description{
Randomly select an element from an object and return a copy of it.
If there is no element, return \code{NULL} or some default value.
This function can be used to sample randomly (with replacement) from a
collection of elements.
}
