% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peek.R
\name{peekleft}
\alias{peekleft}
\alias{peekleft.Deque}
\title{Peek element on left side of an object}
\usage{
peekleft(x)

\method{peekleft}{Deque}(x)
}
\arguments{
\item{x}{any \code{R} object.}

\item{...}{additional arguments to be passed to or from methods.}
}
\value{
For \code{Deque} the first element at the left side of the deque.
}
\description{
Search an element in an object from left to right and return a copy of the
first occurence.  If there is no element, return \code{NULL} or some default value.
}
