% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop.R
\name{pop}
\alias{pop}
\alias{pop_}
\alias{popleft_}
\alias{pop_.Deque}
\alias{popleft_.Deque}
\alias{pop_.Dict}
\alias{pop_.dict.table}
\title{Pop element}
\usage{
pop_(.x, ...)

popleft_(.x, ...)

\method{pop_}{Deque}(.x)

\method{popleft_}{Deque}(.x)

\method{pop_}{Dict}(.x, key)

\method{pop_}{dict.table}(.x, key)
}
\arguments{
\item{.x}{any \code{R} object.}

\item{...}{additional arguments to be passed to or from methods.}

\item{key}{\code{character} name of the value to pop.
the associated key-value pair is deleted and it's value returned.}
}
\value{
For \code{Deque} the first (\code{popleft_}) or last (\code{pop_}) element of the
deque after it was removed.

For \code{Dict} the value associated with the key after the key-value
pair was removed from the dict. If \code{key} is not found, an error is raised.

For \code{dict.table}, returns the column named \code{key} after it was
removed from the dict.table. If column does not exist, an error is raised.
}
\description{
Search and return an element and remove it afterwards from the object.
If the element is not found, signal an error.
}
\details{
All functions work by reference, that is, the original object is altered.
\code{pop_(.x)} tries to access specific values.

\code{popleft_(.x)} pops first element of a \code{Deque}.
}
\examples{
# Deque
d = deque(1, 2, 3)
pop_(d)
popleft_(d)

\dontrun{
pop_(deque())  # pop at empty Deque}

# Dict
d = dict(a = 1, b = 1:3)
pop_(d, "b")
print(d)

\dontrun{
pop_(d, "x")  # key 'x' not in Dict}

# dict.table
dit = dict.table(a = 1:3, b = 4:6)
pop_(dit, "a")
print(dit)
\dontrun{
pop_(dit, "x")  # Column 'x' not in dict.table}
}
\seealso{
\code{\link[=peek]{peek()}}
}
