% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop.R
\name{popleft}
\alias{popleft}
\alias{popleft.Deque}
\title{Pop element from left}
\usage{
popleft(x)

\method{popleft}{Deque}(x)
}
\arguments{
\item{x}{any \code{R} object.}

\item{...}{additional arguments to be passed to or from methods.}
}
\value{
For \code{Deque} the first element at the left side of the deque after it
was removed.
}
\description{
Search an element in an object from left to right and for the first
occurence return a copy and remove it.
If the element is not found, signal an error.
}
