% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev.R
\name{rev}
\alias{rev}
\alias{ref_rev}
\alias{ref_rev.Deque}
\alias{rev.Deque}
\title{Reverse Elements}
\usage{
ref_rev(.x, ...)

\method{ref_rev}{Deque}(x)

\method{rev}{Deque}(x)
}
\arguments{
\item{x}{\code{Deque} object}
}
\value{
For \code{Deque}, an object of class \code{Deque}
}
\description{
\code{rev} provides a reversed version of its argument.
}
\details{
\code{rev} uses copy semantics while \code{ref_rev} works by reference,
that is, it reverse all elements in place.
}
\examples{

d = deque(a = 1, b = 2, 3)
rev(d)
print(d)
ref_rev(d)
print(d)

}
\seealso{
\code{\link[base:rev]{base::rev()}}
}
