% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate.R
\name{rotate}
\alias{rotate}
\alias{rotate.Deque}
\title{Rotate elements}
\usage{
rotate(x, ...)

\method{rotate}{Deque}(x, n = 1L)
}
\arguments{
\item{x}{any \code{R} object.}

\item{...}{additional arguments to be passed to or from methods.}

\item{n}{\code{integer} number of steps to rotate}
}
\value{
For \code{Deque} invisibly returns the \code{\link[=Deque]{Deque()}} object rotated in place.
}
\description{
Rotate all elements \code{n} steps to the right. If n is
negative, rotate to the left.
}
