% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container-deprecated.R
\name{size-deprecated}
\alias{size-deprecated}
\alias{size}
\alias{size.Container}
\title{Container size (deprecated)}
\usage{
size(x)

\method{size}{Container}(x)
}
\value{
\code{integer} size of the \code{Container}, that is, the number of
elements it contains.
}
\description{
This function is deprecated. Use \code{\link[=length]{length()}} instead.
}
