% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container-deprecated.R
\name{sortkey-deprecated}
\alias{sortkey-deprecated}
\alias{sortkey}
\alias{sortkey.Dict}
\title{Sort Dict keys (deprecated)}
\usage{
sortkey(x, decr = FALSE, ...)

\method{sortkey}{Dict}(x, decr = FALSE)
}
\arguments{
\item{x}{any \code{R} object.}

\item{decr}{\code{logical} Should the sort be increasing or decreasing?}

\item{...}{additional arguments to be passed to or from methods.}
}
\value{
For \code{Dict} sorts the elements in place and inivisbly returns the
\code{\link[=Dict]{Dict()}} object.
}
\description{
Re-order elements according to key-order.This function
is deprecated as keys are now always sorted.
}
