% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortkey.R
\name{sortkey}
\alias{sortkey}
\alias{sortkey.Dict}
\title{Sort elements according to their keys}
\usage{
sortkey(x, decreasing = FALSE, ...)

\method{sortkey}{Dict}(x, decreasing = FALSE)
}
\arguments{
\item{x}{any \code{R} object.}

\item{decreasing}{\code{logical} Should the sort be increasing or decreasing?}

\item{...}{additional arguments to be passed to or from methods.}
}
\value{
For \code{Dict} sorts the elements in place and inivisbly returns the
\code{\link[=Dict]{Dict()}} object.
}
\description{
Re-order elements according to key-order.
}
