% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container-deprecated.R
\name{values-deprecated}
\alias{values-deprecated}
\alias{values}
\alias{values.Container}
\alias{values.dict.table}
\title{Get raw values of an object (deprecated)}
\usage{
values(x)

\method{values}{Container}(x)

\method{values}{dict.table}(x)
}
\arguments{
\item{x}{any \code{R} object.}
}
\value{
a copy of all elements as a basic \code{list}.
}
\description{
Get raw values of an object (deprecated)
}
