% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/values.R
\name{values}
\alias{values}
\alias{values.Container}
\title{Get raw values of an object}
\usage{
values(x, ...)

\method{values}{Container}(x)
}
\arguments{
\item{x}{any \code{R} object.}

\item{...}{additional arguments to be passed to or from methods.}
}
\value{
a copy of all elements if possible as \code{atomic} vector otherwise as
a basic \code{list}.
}
\description{
Get raw values of an object
}
