% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add}
\alias{add}
\alias{add.Container}
\alias{add.Dict}
\title{Add element}
\usage{
add(x, ...)

\method{add}{Container}(x, elem)

\method{add}{Dict}(x, key, elem)
}
\arguments{
\item{x}{any \code{R} object.}

\item{...}{additional arguments to be passed to or from methods.}

\item{elem}{some element to be added.}

\item{key}{\code{character} unique identifier of the item.}
}
\value{
For \code{Container}, the container object with the element being added.

For \code{Dict}, the dict object with the key-element pair being added.
}
\description{
Add element
}
