% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add}
\alias{add}
\alias{add.Dict}
\alias{add.dict.table}
\title{Add element}
\usage{
add(x, ...)

\method{add}{Dict}(x, key, value)

\method{add}{dict.table}(x, column, value)
}
\arguments{
\item{x}{any \code{R} object.}

\item{...}{additional arguments to be passed to or from methods.}

\item{key}{\code{character} unique identifier of the item.}

\item{value}{value to be added.}

\item{column}{\code{character} name or \code{numeric} index of column.}

\item{elem}{some element to be added.}
}
\value{
For \code{Dict}, the dict object with the key-element pair being added.

For \code{dict.table}, the dict.table object with the column being added.
}
\description{
Add element
}
