% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{addleft}
\alias{addleft}
\alias{addleft.Deque}
\title{Add element to the left}
\usage{
addleft(x, ...)

\method{addleft}{Deque}(x, elem)
}
\arguments{
\item{x}{any \code{R} object.}

\item{...}{additional arguments to be passed to or from methods.}

\item{elem}{some element to be added.}
}
\value{
For \code{Deque}, the deque object with the element being added to the
left.
}
\description{
Add element to the left
}
