% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear.R
\name{clear}
\alias{clear}
\alias{ref_clear}
\alias{clear.Container}
\alias{ref_clear.Container}
\alias{clear.dict.table}
\alias{ref_clear.dict.table}
\title{Clear a container}
\usage{
clear(x)

ref_clear(x)

\method{clear}{Container}(x)

\method{ref_clear}{Container}(x)

\method{clear}{dict.table}(x)

\method{ref_clear}{dict.table}(x)
}
\arguments{
\item{x}{any \code{R} object.}
}
\value{
For \code{Container}, an object of class \code{Container} (or one of the
respective derived classes).

For \code{dict.table} an object of class \code{dict.table}.
}
\description{
Removes all elements from the container object.
}
\examples{

co = container(1, 2, mean)
clear(co)
co
ref_clear(co)
co

dit = dict.table(a = 1, b = 2)
clear(dit)
dit              # original was not touched
ref_clear(dit)
dit              # original was cleared
}
