% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear.R
\name{clear}
\alias{clear}
\alias{clear.Container}
\alias{clear.dict.table}
\title{Clear an object}
\usage{
clear(x, ...)

\method{clear}{Container}(x)

\method{clear}{dict.table}(x)
}
\arguments{
\item{x}{any \code{R} object.}

\item{...}{additional arguments to be passed to or from methods.}
}
\value{
The cleared object.
}
\description{
Removes all elements from the object.
}
