% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clone.R
\name{clone}
\alias{clone}
\alias{clone.Container}
\title{Clone an object}
\usage{
clone(x, ...)

\method{clone}{Container}(x, deep = FALSE)
}
\arguments{
\item{x}{any \code{R} object.}

\item{...}{additional arguments to be passed to or from methods.}

\item{deep}{\code{logical} if \code{TRUE} a \code{deep} copy otherwise (default) a shallow
copy is performed.}
}
\value{
A copy of the object.
}
\description{
Creates a copy of the object.
}
