% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container-deprecated.R
\name{container-deprecated}
\alias{container-deprecated}
\title{Deprecated Functions in Package container}
\usage{
size(...)  ### -- use  length(...)  instead

sortkey(...)  ### -- `Dict` keys are now always sorted

remove(...)  ### -- use  delete(...)  instead
}
\arguments{
\item{...}{(arguments)}
}
\description{
These functions are provided for backwards-compatibility and
may be defunct as soon as the next release.
}
\seealso{
\code{\link[base:Deprecated]{base::Deprecated()}}
}
