% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container-S3.R
\name{ContainerS3}
\alias{ContainerS3}
\alias{container}
\alias{as.container}
\alias{is.container}
\alias{length.Container}
\alias{as.list.Container}
\title{A sequence Container}
\usage{
container(..., keep_names = FALSE)

as.container(x, ...)

is.container(x)

\method{length}{Container}(x)

\method{as.list}{Container}(x)
}
\arguments{
\item{...}{initial elements put into the \code{Container}.}

\item{keep_names}{\code{logical} if TRUE, keeps names of passed elements.}

\item{x}{any \code{R} object, or an object inheriting from class 'Container' for
the S3 methods.}
}
\value{
\code{\link[=container]{container()}} returns a \link[=container]{Container()} object.

\code{\link[=as.container]{as.container()}} coerces to a container.

\code{\link[=is.container]{is.container()}} returns \code{TRUE} if its argument is a \code{\link[=Container]{Container()}}
and \code{FALSE} otherwise.

\code{length()} returns the length of the internally stored sequence.

\code{as.list()} returns a copy of the internally stored sequence as a
base R list.
}
\description{
A \code{\link[=Container]{Container()}} is a data structure with typical member
functions to insert, delete and access elements from the container
object. Since the \code{\link[=Container]{Container()}} class mainly serves as the
base class for \code{\link[=Deque]{Deque()}}, \code{\link[=Set]{Set()}}, and \code{\link[=Dict]{Dict()}} objects, users are more
likley to use the corresponding \link[=Deque]{deque()}, \link[=Set]{set()}, and \link[=Dict]{dict()} methods to
create objects of the respective derived classes.
}
\details{
For a full list of all container methods see \code{\link[=Container]{Container()}}.
}
\seealso{
\code{\link[=Container]{Container()}}, \code{\link[=deque]{deque()}}, \code{\link[=set]{set()}}, \code{\link[=dict]{dict()}}
}
