% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container-S3.R
\name{ContainerS3op}
\alias{ContainerS3op}
\alias{+.Container}
\title{Binary \code{Container} operators}
\usage{
\method{+}{Container}(c1, c2)
}
\arguments{
\item{c1}{\code{\link[=Container]{Container()}} object}

\item{c2}{\code{\link[=Container]{Container()}} object}
}
\value{
For \code{+} returns a new \code{\link[=Container]{Container()}} object containing all elements
from \code{c1} and \code{c2}.
}
\description{
Binary operators for \code{\link[=Container]{Container()}} objects.
}
