% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{count}
\alias{count}
\alias{count.Deque}
\alias{count.Set}
\title{Count elements}
\usage{
count(x, ...)

\method{count}{Deque}(x, elem)

\method{count}{Set}(x, elem)
}
\arguments{
\item{x}{any \code{R} object.}

\item{...}{additional arguments to be passed to or from methods.}

\item{elem}{element to counted.}
}
\value{
A copy of the object.

\code{integer} times the \code{elem} appears in the object.
}
\description{
Count elements
}
