% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.R
\name{delete}
\alias{delete}
\alias{delete.Dict}
\alias{delete.dict.table}
\title{Delete elements}
\usage{
delete(x, ...)

\method{delete}{Dict}(x, key)

\method{delete}{dict.table}(x, column)
}
\arguments{
\item{x}{any \code{R} object.}

\item{...}{additional arguments to be passed to or from methods.}

\item{key}{\code{character} key of the value to delete. If \code{key} does exists,
the associated key-value pair is deleted, otherwise an error is signaled.}

\item{column}{\code{character} name or \code{numeric} index of column.}
}
\value{
For \code{Dict} the dict object after the key-value pair was removed.

For \code{dict.table} the dict.table object after the column was removed.
}
\description{
Search and remove an element from an object. If the element is not found,
signal an error.
}
