% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.R
\name{delete}
\alias{delete}
\alias{delete.Container}
\alias{delete.Dict}
\title{Delete elements}
\usage{
delete(x, ...)

\method{delete}{Container}(x, elem, right = FALSE)

\method{delete}{Dict}(x, key)
}
\arguments{
\item{x}{any \code{R} object.}

\item{...}{additional arguments to be passed to or from methods.}

\item{elem}{element to be deleted from the object. If not found, an
error is signaled.}

\item{right}{\code{logical} if \code{TRUE}, search from right to left.}

\item{key}{\code{character} key of the value to delete. If \code{key} does exists,
the associated key-value pair is deleted, otherwise an error is signaled.}
}
\value{
For \code{Container} the container object after the element was removed.

For \code{Dict} the dict object after the key-value pair was removed.
}
\description{
Search and remove an element from an object. If the element is not found,
signal an error.
}
