% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deque.R
\name{Deque}
\alias{Deque}
\title{A Deque (double-ended queue) class}
\description{
Deques are a generalization of stacks and queues typically
with methods to add, remove and access elements at both sides of the
underlying data sequence. As such, the \code{\link[=Deque]{Deque()}} can also be used to mimic
both stacks and queues.
}
\details{
This class inherits from class \code{\link[=Container]{Container()}} and extends it by
\code{pop} and \code{peek} methods, element counting, and reverse and rotate
functionality.
}
\seealso{
\code{\link[=Container]{Container()}}, \code{\link[=deque]{deque()}}
}
\author{
Roman Pahl
}
\section{Super classes}{
\code{\link[container:Iterable]{container::Iterable}} -> \code{\link[container:Container]{container::Container}} -> \code{Deque}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-addleft}{\code{Deque$addleft()}}
\item \href{#method-count}{\code{Deque$count()}}
\item \href{#method-peek}{\code{Deque$peek()}}
\item \href{#method-peekleft}{\code{Deque$peekleft()}}
\item \href{#method-pop}{\code{Deque$pop()}}
\item \href{#method-popleft}{\code{Deque$popleft()}}
\item \href{#method-reverse}{\code{Deque$reverse()}}
\item \href{#method-rotate}{\code{Deque$rotate()}}
\item \href{#method-clone}{\code{Deque$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="container" data-topic="Iterable" data-id="iter">}\href{../../container/html/Iterable.html#method-iter}{\code{container::Iterable$iter()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="container" data-topic="Container" data-id="add">}\href{../../container/html/Container.html#method-add}{\code{container::Container$add()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="container" data-topic="Container" data-id="apply">}\href{../../container/html/Container.html#method-apply}{\code{container::Container$apply()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="container" data-topic="Container" data-id="clear">}\href{../../container/html/Container.html#method-clear}{\code{container::Container$clear()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="container" data-topic="Container" data-id="discard">}\href{../../container/html/Container.html#method-discard}{\code{container::Container$discard()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="container" data-topic="Container" data-id="empty">}\href{../../container/html/Container.html#method-empty}{\code{container::Container$empty()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="container" data-topic="Container" data-id="has">}\href{../../container/html/Container.html#method-has}{\code{container::Container$has()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="container" data-topic="Container" data-id="initialize">}\href{../../container/html/Container.html#method-initialize}{\code{container::Container$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="container" data-topic="Container" data-id="print">}\href{../../container/html/Container.html#method-print}{\code{container::Container$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="container" data-topic="Container" data-id="remove">}\href{../../container/html/Container.html#method-remove}{\code{container::Container$remove()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="container" data-topic="Container" data-id="size">}\href{../../container/html/Container.html#method-size}{\code{container::Container$size()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="container" data-topic="Container" data-id="type">}\href{../../container/html/Container.html#method-type}{\code{container::Container$type()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="container" data-topic="Container" data-id="values">}\href{../../container/html/Container.html#method-values}{\code{container::Container$values()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addleft"></a>}}
\if{latex}{\out{\hypertarget{method-addleft}{}}}
\subsection{Method \code{addleft()}}{
Add elem to left side of the \code{Deque}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Deque$addleft(elem)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{elem}}{element to be added.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisibly returns the \code{Deque()} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-count"></a>}}
\if{latex}{\out{\hypertarget{method-count}{}}}
\subsection{Method \code{count()}}{
Count number of elem occurences.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Deque$count(elem)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{elem}}{element to be counted.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{integer} number of \code{elem} occurences in the \code{\link[=Deque]{Deque()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-peek"></a>}}
\if{latex}{\out{\hypertarget{method-peek}{}}}
\subsection{Method \code{peek()}}{
Peek at last element on the right without removing it.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Deque$peek()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
element 'peeked' on the right
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-peekleft"></a>}}
\if{latex}{\out{\hypertarget{method-peekleft}{}}}
\subsection{Method \code{peekleft()}}{
Peek at first element on the left without removing it.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Deque$peekleft()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
element 'peeked' on the left
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pop"></a>}}
\if{latex}{\out{\hypertarget{method-pop}{}}}
\subsection{Method \code{pop()}}{
Remove and return element from the right side of the \code{\link[=Deque]{Deque()}}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Deque$pop()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
element 'popped' from the right side of the \code{\link[=Deque]{Deque()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-popleft"></a>}}
\if{latex}{\out{\hypertarget{method-popleft}{}}}
\subsection{Method \code{popleft()}}{
Remove and return element from the right side of the \code{\link[=Deque]{Deque()}}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Deque$popleft()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
element 'popped' from the left side of the \code{\link[=Deque]{Deque()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reverse"></a>}}
\if{latex}{\out{\hypertarget{method-reverse}{}}}
\subsection{Method \code{reverse()}}{
Reverse all elements of the \code{\link[=Deque]{Deque()}} in-place.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Deque$reverse()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
invisibly returns the \code{Deque()} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rotate"></a>}}
\if{latex}{\out{\hypertarget{method-rotate}{}}}
\subsection{Method \code{rotate()}}{
Rotate all elements \code{n} steps to the right. If n is
negative, rotate to the left.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Deque$rotate(n = 1L)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{\code{integer} number of steps to rotate}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisibly returns the \code{Deque()} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Deque$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
