% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dequeS3.R
\name{dequeS3}
\alias{dequeS3}
\alias{deque}
\alias{as.deque}
\alias{as.deque.Container}
\alias{is.deque}
\alias{rev.Deque}
\title{Deque (double-ended queue)}
\usage{
deque(..., keep_names = FALSE)

as.deque(x, ...)

\method{as.deque}{Container}(x, ...)

is.deque(x)

\method{rev}{Deque}(x)
}
\arguments{
\item{...}{initial elements put into the \code{Deque}.}

\item{keep_names}{\code{logical} if TRUE, keeps names of passed elements.}

\item{x}{any \code{R} object, or an object inheriting from class 'Deque' or
'Container' for the corresponding S3 methods.}
}
\value{
\code{\link[=deque]{deque()}} returns a \link[=deque]{Deque()} object.

\code{\link[=as.deque]{as.deque()}} coerces to a deque.

\code{\link[=is.deque]{is.deque()}} returns \code{TRUE} if its argument is a \code{\link[=Deque]{Deque()}}
and \code{FALSE} otherwise.

\code{rev} reverses all elements of the deque in place and invisibly
returns the \code{\link[=Deque]{Deque()}} object.
}
\description{
Deques are a generalization of stacks and queues typically
with methods to add, remove and access elements at both sides of the
underlying data sequence. As such, the \code{\link[=deque]{deque()}} can also be used to mimic
both stacks and queues.
}
\details{
For a full list of all deque methods see \code{\link[=Deque]{Deque()}}.
}
\seealso{
\code{\link[=Deque]{Deque()}}, \code{\link[=container]{container()}}
}
