% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deque-S3.R
\name{dequeS3}
\alias{dequeS3}
\alias{deque}
\alias{as.deque}
\alias{is.deque}
\alias{addleft}
\alias{count}
\alias{peek}
\alias{peekleft}
\alias{pop}
\alias{popleft}
\alias{reverse}
\alias{rotate}
\alias{addleft.Deque}
\alias{count.Deque}
\alias{peek.Deque}
\alias{peekleft.Deque}
\alias{pop.Deque}
\alias{popleft.Deque}
\alias{reverse.Deque}
\alias{rotate.Deque}
\title{Deque (double-ended queue) S3 interface}
\usage{
deque(...)

as.deque(x)

is.deque(x)

addleft(x, ...)

count(x, ...)

peek(x, ...)

peekleft(x)

pop(x, ...)

popleft(x)

reverse(x)

rotate(x, ...)

\method{addleft}{Deque}(x, elem, ...)

\method{count}{Deque}(x, elem, ...)

\method{peek}{Deque}(x, ...)

\method{peekleft}{Deque}(x)

\method{pop}{Deque}(x, ...)

\method{popleft}{Deque}(x)

\method{reverse}{Deque}(x)

\method{rotate}{Deque}(x, n = 1L, ...)
}
\arguments{
\item{...}{further arguments depending on the method.}

\item{x}{a vector of 'any' type}

\item{elem}{some element of the deque}

\item{n}{\code{integer} number of steps to rotate}
}
\description{
Deques are a generalization of stacks and queues typically
with methods to add, remove and access elements at both sides of the
underlying data sequence. As such, the \code{\link[=deque]{deque()}} can also be used to mimic
both stacks and queues.
}
\details{
For a detailed documentation of all methods see \code{\link[=Deque]{Deque()}}.
}
\seealso{
\code{\link[=Container]{Container()}}, \code{\link[=+.Deque]{+.Deque()}}
}
