% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DequeS3.R, R/GroupGenericMath.R,
%   R/GroupGenericSummary.R, R/Ops-arithmetic.R, R/add.R, R/peek.R, R/pop.R,
%   R/rotate.R
\name{DequeS3}
\alias{DequeS3}
\alias{deque}
\alias{as.deque}
\alias{is.deque}
\alias{rev.Deque}
\alias{OpsMath}
\alias{OpsSummary}
\alias{OpsArith}
\alias{DequeS3methods}
\alias{addleft.Deque}
\alias{peek.Deque}
\alias{pop.Deque}
\alias{rotate.Deque}
\title{Deque (double-ended queue)}
\usage{
deque(...)

as.deque(x)

is.deque(x)

\method{rev}{Deque}(x)

## Deque S3 methods

addleft(.x, ...)
addleft_(.x, ...)

peek(x, default)
peekleft(x, default)

pop_(.x)
popleft_(.x)

rotate(x, n)
}
\arguments{
\item{...}{initial elements put into the \code{Deque}.}

\item{x}{\code{R} object of \code{ANY} type for \code{\link[=as.deque]{as.deque()}} and \code{\link[=is.deque]{is.deque()}}
or of class \code{Deque} for the \code{S3} methods.}

\item{.x}{\verb{object of class }Deque`}

\item{default}{value to be returned if peeked value does not exist.}

\item{n}{\code{integer} number of steps to rotate}
}
\description{
Deques are a generalization of stacks and queues typically
with methods to add, remove and access elements at both sides of the
underlying data sequence. As such, the \code{\link[=deque]{deque()}} can also be used to mimic
both stacks and queues.
}
\details{
Methods that alter \code{Deque} objects usually come in two versions
providing either copy or reference semantics where the latter are visible
by an underscore appended to the standard function name, for example,
\verb{[addleft()]} and \verb{[addleft_()]}.
\subsection{Methods}{
}

\itemize{
\item \code{deque(...)} initializes and returns an object of class \code{Deque}
}

\itemize{
\item \code{as.deque(x)} coerces \code{x} to a deque.
}

\itemize{
\item \code{is.deque(x)} returns \code{TRUE} if \code{x} is of class \code{Deque}
and \code{FALSE} otherwise.
}

\itemize{
\item \code{rev(x)} reverses all elements of the deque in place and invisibly
returns the \code{\link[=Deque]{Deque()}} object.
}

\itemize{
\item \code{x - y} element-wise removes all items of \code{y} from \code{x}, given
the element was contained in \code{x}.
}

\itemize{
\item \code{addleft(.x, ...)} adds (possibly named) elements to left side of \code{.x}.
\item \code{addleft_(.x, ...)} same as \code{addleft(.x, ...)} but adds by reference.
}

\itemize{
\item \code{peek(x, default = NULL)} peek at last element. If \code{x} is empty, return
\code{default}.
\item \code{peekleft(x, default = NULL)} peek at first element. If \code{x} is empty,
return \code{default}.
}

\itemize{
\item \code{pop_(.x)} pop last element. If \code{.x} is empty, an error is given.
\item \code{popleft_(.x)} pop first element. If \code{.x} is empty, an error is given.
}

\itemize{
\item \code{rotate(x, n)} rotate all elements \code{n} steps to the right, If \code{n} is
negative, rotate to the left.
}
}
\examples{
d = deque(1, 2, s = "a", v = 1:3)
is.deque(d)
print(d)
length(d)
names(d)
as.list(d)
rev(d)

l = list(0, 1)
d2 = as.deque(l)
d + d2
c(d, d2) # same as d + d2
d2 + d
d - d2
c(d2, d) # same as d2 + d
d2 - d

# Math
set.seed(123)
d = as.deque(rnorm(3))
abs(d)
cumsum(d)
round(d)
exp(d)

# Summary
range(d)
min(d)
max(d)

d1 = deque(1, 1:2)
d2 = deque(2, 1:2)
d1 + d2     # same as c(d1, d2)
d2 + d1     # same as c(d2, d1)
d1 - d2
d2 - d1
d1 - d1


d = deque(0)
add(d, a = 1, b = 2)         # |0, a = 1, b = 2|
addleft(d, a = 1, b = 2)     # |b = 2, a = 1, 0|

d = deque(1, 2, 3)
peek(d)
peekleft(d)
peek(deque())
peek(deque(), default = 0)
peekleft(deque(), default = 0)
d = deque(1, 2, 3)
pop_(d)
print(d)
popleft_(d)
print(d)

\dontrun{
pop_(deque())  # pop at empty Deque}

d = deque(1, 2, 3, 4)
rotate(d)
rotate(d, n = 2)

}
\seealso{
See \code{\link[=container]{container()}} for all inherited methods. For the full class
documentation see \code{\link[=Deque]{Deque()}} and it's superclass \link[=container]{Container()}.
}
