% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deque-S3.R
\name{dequeS3binOp}
\alias{dequeS3binOp}
\alias{+.Deque}
\title{Binary \code{Deque} operators}
\usage{
\method{+}{Deque}(x1, x2)
}
\arguments{
\item{x1}{primitive or \code{\link[=Deque]{Deque()}} object}

\item{x2}{primitive or \code{\link[=Deque]{Deque()}} object}
}
\value{
For \code{+} if left side is a \code{\link[=Deque]{Deque()}}, adds element to right side of
the \code{\link[=Deque]{Deque()}}. If right side is a \code{\link[=Deque]{Deque()}}, adds element to left side of
the \code{\link[=Deque]{Deque()}}.
}
\description{
Binary operators for \code{\link[=Deque]{Deque()}} objects.
}
