% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict.frame-S3.R
\name{dict.frameS3replace}
\alias{dict.frameS3replace}
\alias{[[.Dict.frame}
\alias{[.Dict.frame}
\title{Extract or replace parts of a \code{Dict.frame}}
\usage{
\method{[[}{Dict.frame}(x, i, j, default = NULL)

\method{[}{Dict.frame}(x, i, j, default = NULL)
}
\arguments{
\item{x}{\code{\link[=Dict.frame]{Dict.frame()}} object.}

\item{i, }{j \code{integer} or \code{character} elements to extract or replace.}

\item{default}{A suitable default value.}
}
\value{
For \code{[[} returns the value at row \code{i} and column \code{j} or \code{default} if
the column does not exist. If \code{i} is not specified (or only one index
given), the full column is returned, if it exists or \code{default} if not.

For \code{[} returns a \code{Dict.frame} containing the extracted values.
}
\description{
Access and assignment operators for \code{\link[=Dict.frame]{Dict.frame()}} objects.
}
