% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict.frame-S3.R
\name{dict.frameS3replace}
\alias{dict.frameS3replace}
\alias{[[.Dict.frame}
\alias{[.Dict.frame}
\alias{[[<-.Dict.frame}
\alias{[<-.Dict.frame}
\title{Extract or replace parts of a \code{Dict.frame}}
\usage{
\method{[[}{Dict.frame}(x, i, j, default = NULL)

\method{[}{Dict.frame}(x, i, j, default = NULL)

\method{[[}{Dict.frame}(x, i, j, add = FALSE) <- value

\method{[}{Dict.frame}(x, i, j, add = FALSE) <- value
}
\arguments{
\item{x}{\code{\link[=Dict.frame]{Dict.frame()}} object.}

\item{i}{\code{integer} row indices of elements to extract or replace.}

\item{j}{\code{integer} or \code{character} column indices of elements to extract
or replace.}

\item{default}{A suitable default value.}

\item{add}{\code{logical} If only the column index was set
(i.e. \code{x[i]} or \code{x[, j]}), #' and \code{add} set to \code{TRUE} and the column index
is of type \code{character} then the
the value(s) are added as a new column to the \code{dict.frame}. In all other
cases or if \code{add == FALSE}, an out of bounds error is signaled.}

\item{value}{A suitable replacement value.}
}
\value{
For \code{[[} returns the value at row \code{i} and column \code{j}. If the column
is specified as an integer and does not exist, an out of bounds error is
raised. If specified as character, \code{NULL} is returned unless \code{default} was
specified, in which case the \code{default} value is returned.
If \code{i} is not specified (or only one index given), the full column is
returned, again, following the above rules in case of missing column.

For '[' returns the values at rows \code{i} and columns \code{j}. For any of
the columns being specified as an integer that do not exist, an out of
bounds error is raised. If specified as character, an error is raised
unless \code{default} was specified, in which case the \code{default} value is
returned. In this case, the \code{default} values must be a multiple of the
number of rows. Note that it is possible to use a mixed
\code{integer}/\code{character} specification of columns passed as a list.
If \code{i} is not specified (or only one index given), the full columns are
returned, again, following the above rules in case of missing columns.

For '[[<-' replaces the value at position i, j. If a new column
is generated (see option \code{add}), the value is replicated if needed. Note
that for this the value has to be a multiple of the number of rows. In
contrast to \code{\link[base:data.frame]{base::data.frame()}}s, numerical row or column indices that
exceed the dimensions of the \code{\link[=dict.frame]{dict.frame()}} will always stop with an error,
that is, new rows or columns are never substituted. The only way to attach
a new column is to use a single character index and explicitly stating the
intention to add the column, i.e., \code{x[["A", add = TRUE]] <- 1}.

For '[<-' replaces the value(s) at positions i, j. In
contrast to \code{\link[base:data.frame]{base::data.frame()}}s, numerical row or column indices that
exceed the dimensions of the \code{\link[=dict.frame]{dict.frame()}} will always stop with an error,
that is, new rows or columns are never substituted. The only way to attach
new columns is to use character indices, i.e., \code{x["A", add = TRUE] <- 1} or
\code{x[, c("A", "C"), add = TRUE] <- 9} .
}
\description{
Access and assignment operators for \code{\link[=Dict.frame]{Dict.frame()}} objects.
}
