% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict.table-S3.R
\name{dict.tableS3}
\alias{dict.tableS3}
\alias{dict.table}
\alias{as.dict.table}
\alias{as.dict.table.data.table}
\alias{as.dict.table.default}
\alias{is.dict.table}
\alias{add.dict.table}
\alias{clear.dict.table}
\alias{delete.dict.table}
\alias{discard.dict.table}
\alias{empty.dict.table}
\alias{getval.dict.table}
\alias{has.dict.table}
\alias{keys.dict.table}
\alias{peek.dict.table}
\alias{pop.dict.table}
\alias{popitem.dict.table}
\alias{print.dict.table}
\alias{rbind.dict.table}
\alias{rename.dict.table}
\alias{setval.dict.table}
\alias{size.dict.table}
\alias{sortkey.dict.table}
\alias{values.dict.table}
\title{dict.table}
\usage{
\method{dict}{table}(...)

as.dict.table(x, ...)

\method{as.dict}{table.data.table}(x, copy = TRUE, ...)

\method{as.dict}{table.default}(x)

is.dict.table(x)

\method{add}{dict.table}(x, key, value)

\method{clear}{dict.table}(x)

\method{delete}{dict.table}(x, key)

\method{discard}{dict.table}(x, key)

\method{empty}{dict.table}(x)

\method{getval}{dict.table}(x, key)

\method{has}{dict.table}(x, key)

keys.dict.table(x)

\method{peek}{dict.table}(x, key, default = NULL)

\method{pop}{dict.table}(x, key)

\method{popitem}{dict.table}(x)

\method{print}{dict.table}(x, ...)

\method{rbind}{dict.table}(x, ...)

\method{rename}{dict.table}(x, old, new)

\method{setval}{dict.table}(x, key, value, add = FALSE)

size.dict.table(x)

\method{sortkey}{dict.table}(x, decr = FALSE)

\method{values}{dict.table}(x)
}
\arguments{
\item{...}{arguments of the form tag = value or a named list}

\item{x}{a \code{dict.table} object.}
}
\description{
The \code{\link[=dict.table]{dict.table()}} is a mix of a dictionary and a
\code{data.table}. In particular, it is a dictionary thereby inheriting all
\code{\link[=Dict]{Dict()}} methods, but where each element has the same length. In contrast to
a \code{\link[base:data.frame]{base::data.frame()}}, a \code{\link[=dict.table]{dict.table()}} can contain arbitrary complex
objects.
}
\details{
For a detailed documentation of all methods see \code{\link[=Dict.table]{Dict.table()}}.
}
\seealso{
\code{\link[=Dict]{Dict()}}, \code{\link[=data.table:data.table]{data.table:data.table()}}, [\verb{[[<-.dict.table()}],
[\verb{[[.dict.table()}], [\verb{[<-.dict.table()}], [\verb{[.dict.table()}]
}
