% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict-S3.R
\name{dictS3}
\alias{dictS3}
\alias{dict}
\alias{dict.data.frame}
\alias{dict.default}
\alias{as.dict}
\alias{is.dict}
\alias{getval}
\alias{keys}
\alias{popitem}
\alias{setval}
\alias{rename}
\alias{sortkey}
\alias{getval.Dict}
\alias{keys.Dict}
\alias{popitem.Dict}
\alias{rename.Dict}
\alias{setval.Dict}
\alias{sortkey.Dict}
\alias{update.Dict}
\title{Dict S3 interface}
\usage{
dict(...)

\method{dict}{data.frame}(...)

\method{dict}{default}(...)

as.dict(x)

is.dict(x)

getval(x, ...)

keys(x)

popitem(x)

setval(x, ...)

rename(x, ...)

sortkey(x, ...)

\method{getval}{Dict}(x, key, ...)

\method{keys}{Dict}(x)

\method{popitem}{Dict}(x)

\method{rename}{Dict}(x, old, new, ...)

\method{setval}{Dict}(x, key, value, add = FALSE, ...)

\method{sortkey}{Dict}(x, decr = FALSE, ...)

\method{update}{Dict}(object, other, ...)
}
\arguments{
\item{...}{further arguments depending on the method.}

\item{x}{a named vector (or list) of 'any' type including
\code{\link[base:data.frame]{base::data.frame()}}s.}

\item{add}{\code{logical} if FALSE and \code{key} is not in dictionary, an error is
given, otherwise if \code{TRUE}, the key-value pair would be added if the \code{key}
is not yet in the dict.}

\item{decr}{\code{logical} if \code{TRUE} the elements are sorted decreasingly}
}
\description{
The \code{\link[=dict]{dict()}} resembles Python's dict type, and is implemented
as a specialized associative \code{\link[=container]{container()}} thus sharing all \code{\link[=container]{container()}}
methods with some of them being overridden to account for the associative
key-value pair semantic.
}
\details{
For a detailed documentation of all methods see \code{\link[=Dict]{Dict()}}.
}
\seealso{
\code{\link[=Container]{Container()}}, \code{\link[=+.Dict]{+.Dict()}}, \code{\link[=-.Dict]{-.Dict()}}, [\verb{[[<-.Dict()}],
[\verb{[[.Dict()}], [\verb{[<-.Dict()}], [\verb{[.Dict()}]
}
