% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict-S3.R
\name{dictS3}
\alias{dictS3}
\alias{getval}
\alias{keys}
\alias{popitem}
\alias{setval}
\alias{sortkey}
\alias{dict}
\alias{as.dict}
\alias{is.dict}
\alias{add.Dict}
\alias{discard.Dict}
\alias{getval.Dict}
\alias{has.Dict}
\alias{keys.Dict}
\alias{peek.Dict}
\alias{pop.Dict}
\alias{popitem.Dict}
\alias{delete.Dict}
\alias{setval.Dict}
\alias{sortkey.Dict}
\title{Dict S3 interface}
\usage{
getval(x, ...)

keys(x)

popitem(x)

setval(x, ...)

sortkey(x, ...)

dict(x = list())

as.dict(x)

is.dict(x)

\method{add}{Dict}(x, key, value, ...)

\method{discard}{Dict}(x, key, ...)

\method{getval}{Dict}(x, key, ...)

\method{has}{Dict}(x, key, ...)

\method{keys}{Dict}(x)

\method{peek}{Dict}(x, key, default = NULL, ...)

\method{pop}{Dict}(x, key, ...)

\method{popitem}{Dict}(x)

\method{delete}{Dict}(x, key, ...)

\method{setval}{Dict}(x, key, value, add = FALSE, ...)

\method{sortkey}{Dict}(x, decr = FALSE, ...)
}
\arguments{
\item{x}{a named vector (or list) of 'any' type including
\code{\link[base:data.frame]{base::data.frame()}}s.}

\item{...}{further arguments depending on the method.}

\item{key}{\code{character} unique identifier to access the key-value pair.}

\item{value}{value to add to the dictionary.}

\item{default}{default value to return if \code{key} not in the dictionary}

\item{add}{\code{logical} if FALSE and \code{key} is not in dictionary, an error is
given, otherwise if \code{TRUE}, the key-value pair would be added if the \code{key}
is not yet in the dict.}

\item{decr}{\code{logical} if \code{TRUE} the elements are sorted decreasingly}
}
\description{
The \code{\link[=dict]{dict()}} resembles Python's dict type, and is implemented
as a specialized associative \code{\link[=container]{container()}} thus sharing all \code{\link[=container]{container()}}
methods with some of them being overridden to account for the associative
key-value pair semantic.
}
\details{
For a detailed documentation of all methods see \code{\link[=Dict]{Dict()}}.
}
\seealso{
\code{\link[=Container]{Container()}}, \code{\link[=+.Dict]{+.Dict()}}, \code{\link[=-.Dict]{-.Dict()}}, [\verb{[[<-.Dict()}],
[\verb{[[.Dict()}], [\verb{[<-.Dict()}], [\verb{[.Dict()}]
}
