% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-DictS3.R, R/GroupGenericMath.R,
%   R/GroupGenericSummary.R, R/Ops-arithmetic.R, R/Ops-logic.R, R/add.R,
%   R/rename.R, R/replace.R, R/update.R
\name{DictS3}
\alias{DictS3}
\alias{dict}
\alias{as.dict}
\alias{is.dict}
\alias{DictS3methods}
\alias{+.Dict}
\alias{-.Dict}
\alias{&.Dict}
\alias{|.Dict}
\title{Dict methods}
\usage{
dict(...)

as.dict(x)

is.dict(x)

\method{+}{Dict}(x, y)

\method{-}{Dict}(x, y)

\method{&}{Dict}(x, y)

\method{|}{Dict}(x, y)

## More dict S3 methods

rename(.x, old, new)
ref_rename(.x, old, new)

replace(.x, old, new)
ref_replace(.x, old, new)

update(x, other)
ref_update(x, other)
}
\arguments{
\item{...}{elements put into the \code{Dict}.}

\item{x, y}{\code{R} object of \code{ANY} type for \code{\link[=as.dict]{as.dict()}} and \code{\link[=is.dict]{is.dict()}}
or of class \code{Dict} for the \code{S3} methods.}

\item{key}{\code{character} name of key.}

\item{old}{old element to be found and replaced.}

\item{new}{the new element replacing the old one.}

\item{other}{another \code{Dict} object}
}
\description{
The \code{\link[=dict]{dict()}} resembles Python's dict type, and is implemented
as a specialized associative \code{\link[=Container]{Container()}} thus sharing all \link[=Container]{container()}
methods with some of them being overridden to account for the associative
key-value pair semantic.
}
\details{
Internally, all key-value pairs are stored in a hash-table and the
elements are sorted lexicographically by their keys.
Methods that alter \code{Dict} objects usually come in two versions
providing either copy or reference semantics where the latter start with
\code{'ref_'} to note the reference semantic, for example, \code{add()} and \code{ref_add()}.
\subsection{Methods}{

Description of \code{dict} methods:
}

\itemize{
\item \code{dict(...)} initializes and returns an object of class \code{Dict}
}

\itemize{
\item \code{as.dict(x)} coerces \code{x} to a dictionary
}

\itemize{
\item \code{is.dict(x)} returns \code{TRUE} if \code{x} is of class \code{Dict}
and \code{FALSE} otherwise.
}

\itemize{
\item \code{x + y} combines \code{x} and \code{y} into a new dict by updating \code{x}
by \code{y} (see also \verb{[update()]}).
}

\itemize{
\item \code{x - y} removes all keys from \code{x} that appear in \code{y}.
}

\itemize{
\item \code{x} \code{&} \code{y} returns a copy of \code{x} keeping only the keys that
are common in both (key intersection), that is, all keys in \code{x} that do not
exist in \code{y} are removed.
}

\itemize{
\item \code{x} \code{|} \code{y} returns a copy of \code{x} extended by all elements of
\code{y} that are stored at keys (or names) that do not exist in \code{x}, thereby
combining the keys of both objects (set union of keys).
}

\itemize{
\item \code{rename(.x, old, new)} and \code{ref_rename(.x, old, new)} rename one or more keys
from \code{old} to \code{new}, respectively, by copy and in place (i.e. by reference).
}

\itemize{
\item \code{replace(.x, old, new)} and \code{ref_replace(.x, old)} try to find element \code{old}
and replace it with element \code{new}. If \code{old} does not exist, an error is
raised.
}

\itemize{
\item \code{update(x, other)} and \code{ref_update(x, other)} adds elements of \code{other} dict
for keys not yet in \code{x} and replaces the values of existing keys.
}
}
\examples{
d = dict(a = 1, b = "one", f = mean, na = NA)
print(d)
names(d)
as.list(d)

\dontrun{
dict(a = 1, 2)   # all elements must be named}

# Coercion
as.dict(list(A = 1:3, B = "b"))
as.dict(c(x = 1, y = "x", z = 2 + 3))

# Math
set.seed(123)
d = dict(a = rnorm(1), b = rnorm(1))
abs(d)
cumsum(d)
round(d)
exp(d)

# Summary
range(d)
min(d)
max(d)

d1 = dict(a = 1, b = list(1, 2))
d2 = dict(a = 2, b = list(1, 2))
d1 + d2      # same as update(d, d2)
d2 + d1      # same as update(d2, d)
\dontrun{
c(d1, d2)    # duplicated keys are not allowed for Dict}

d - d2
d2 - d
d - d

d1 = dict(a = 1, b = 2)
d2 = dict(a = 10, x = 4)
d1 & d2      # {a = 1}

d1 | d2      # {a = 1, b = 2, x = 4}

co = container(a = 1, b = 2, 3)
rename(co, c("a", "b"), c("a1", "y"))
print(co)
ref_rename(co, c("a", "b"), c("a1", "y"))
print(co)

d = dict(a = 1, b = "z")
replace(d, 1, 1:5)
replace(d, "z", "a")

\dontrun{
replace(d, "a", 2)              # old element ("a") is not in Dict}


d1 = dict(a = 1, b = 2)
d2 = dict(       b = 0, c = 3)
update(d1, d2)  # {a = 1, b = 0, c = 3}
update(d2, d1)  # {a = 1, b = 2, c = 3}
}
\seealso{
See \code{\link[=container]{container()}} for all inherited methods. For the full class
documentation see \code{\link[=Dict]{Dict()}} and it's superclass \link[=container]{Container()}.
}
