% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DictS3.R
\name{dictS3}
\alias{dictS3}
\alias{dict}
\alias{as.dict}
\alias{is.dict}
\alias{keys}
\alias{names.Dict}
\title{Dict methods}
\usage{
dict(...)

as.dict(x)

is.dict(x)

keys(x)

\method{names}{Dict}(x)
}
\arguments{
\item{...}{initial elements put into the \code{Dict}.}

\item{x}{\code{R} object of \code{ANY} type for \code{\link[=as.dict]{as.dict()}} and \code{\link[=is.dict]{is.dict()}}
or of class \code{Dict} for the \code{S3} methods.}
}
\value{
\code{keys()} returns a \code{character} vector of all the dict's keys.

\code{names()} returns the same as \code{\link[=keys]{keys()}}.
}
\description{
The \code{\link[=dict]{dict()}} resembles Python's dict type, and is implemented
as a specialized associative \code{\link[=Container]{Container()}} thus sharing all \link[=Container]{container()}
methods with some of them being overridden to account for the associative
key-value pair semantic.
}
\details{
For a full list of all dict methods see \code{\link[=Dict]{Dict()}}.

While the \code{\link[=Dict]{Dict()}} class is based on the \code{R6} framework and
provides reference semantics, the presented methods provide an \code{S3}
interface with copy semantics. Note that any \code{S3} methods defined for the
\code{Container} class also work with \code{Deque} objects, even if not overwritten
explicitly.
\subsection{Methods}{
}

\itemize{
\item \code{dict(...)} initializes and returns an object of class \code{Dict}
}

\itemize{
\item \code{as.dict(x)} coerces \code{x} to a deque.
}

\itemize{
\item \code{is.dict(x)} returns \code{TRUE} if \code{x} is of class \code{Deque}
and \code{FALSE} otherwise.
}
}
\examples{
dict(a = 1, b = "one", f = mean)

d1 = dict(a = 1, b = 2)
mode(values(d1))                     # 'list'
d2 = dict(c(a = 1, b = 2))
mode(values(d2))                     # 'numeric'

\dontrun{
dict(a = 1, 2)                       # all elements must be named}

# Coercion
as.dict(list(A = 1:3, B = "b"))
as.dict(c(x = 1, y = "x", z = 2 + 3))

d = dict(x = 1, y = 2)
keys(d)
names(d)
}
\seealso{
See \code{\link[=container]{container()}} for all inherited methods. For the full class
documentation see \code{\link[=Dict]{Dict()}} and it's superclass \link[=container]{Container()}.
}
