% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict-S3.R
\name{dictS3}
\alias{dictS3}
\alias{getval}
\alias{keys}
\alias{peek}
\alias{pop}
\alias{popitem}
\alias{setval}
\alias{sortkey}
\alias{dict}
\alias{as.dict}
\alias{is.dict}
\alias{add.Dict}
\alias{discard.Dict}
\alias{getval.Dict}
\alias{has.Dict}
\alias{keys.Dict}
\alias{peek.Dict}
\alias{pop.Dict}
\alias{popitem.Dict}
\alias{remove.Dict}
\alias{setval.Dict}
\alias{sortkey.Dict}
\title{Dict S3 interface}
\usage{
getval(x, ...)

keys(x)

peek(x, ...)

pop(x, ...)

popitem(x)

setval(x, ...)

sortkey(x, ...)

dict(x = list())

as.dict(x)

is.dict(x)

\method{add}{Dict}(x, key, value, ...)

\method{discard}{Dict}(x, key, ...)

\method{getval}{Dict}(x, key, ...)

\method{has}{Dict}(x, key, ...)

\method{keys}{Dict}(x)

\method{peek}{Dict}(x, key, default = NULL, ...)

\method{pop}{Dict}(x, key, ...)

\method{popitem}{Dict}(x)

\method{remove}{Dict}(x, key, ...)

\method{setval}{Dict}(x, key, value, add = FALSE, ...)

\method{sortkey}{Dict}(x, decr = FALSE, ...)
}
\description{
The \code{\link[=dict]{dict()}} resembles Python's dict type, and is implemented
as a specialized associative \code{\link[=container]{container()}} thus sharing all \code{\link[=container]{container()}}
methods with some of them being overridden to account for the associative
key-value pair semantic.
}
\details{
For a detailed documentation of all methods see \code{\link[=Dict]{Dict()}}.
}
\seealso{
\code{\link[=Container]{Container()}}, \code{\link[=+.Dict]{+.Dict()}}, \code{\link[=-.Dict]{-.Dict()}}, [\verb{[[<-.Dict()}],
[\verb{[[.Dict()}], [\verb{[<-.Dict()}], [\verb{[.Dict()}]
}
