% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict-S3.R
\name{dictS3}
\alias{dictS3}
\alias{dict}
\alias{dict.data.frame}
\alias{as.dict}
\alias{as.dict.data.frame}
\alias{is.dict}
\alias{keys}
\alias{names.Dict}
\alias{update.Dict}
\title{Dict S3 interface}
\usage{
dict(...)

\method{dict}{data.frame}(...)

as.dict(x)

\method{as.dict}{data.frame}(x)

is.dict(x)

keys(x)

\method{names}{Dict}(x)

\method{update}{Dict}(object, other, ...)
}
\arguments{
\item{...}{initial elements put into the \code{Dict}.}

\item{x}{any \code{R} object, or an object inheriting from class 'Dict' or
'Container'.}
}
\value{
\code{\link[=dict]{dict()}} returns a \link[=dict]{Dict()} object.

\code{\link[=as.dict]{as.dict()}} coerces to a dict.

\code{\link[=is.dict]{is.dict()}} returns \code{TRUE} if its argument is a \code{\link[=Dict]{Dict()}} and \code{FALSE}
otherwise.

\code{keys()} returns a \code{character} vector of all the dict's keys.

\code{names()} returns the same as \code{\link[=keys]{keys()}}.
}
\description{
The \code{\link[=dict]{dict()}} resembles Python's dict type, and is implemented
as a specialized associative \code{\link[=container]{container()}} thus sharing all \code{\link[=container]{container()}}
methods with some of them being overridden to account for the associative
key-value pair semantic.
}
\details{
For a full list of all dict methods see \code{\link[=Dict]{Dict()}}.
}
\seealso{
\code{\link[=Dict]{Dict()}}, \code{\link[=container]{container()}}
}
