% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DictS3.R, R/add.R
\name{DictS3}
\alias{DictS3}
\alias{dict}
\alias{as.dict}
\alias{is.dict}
\alias{add.Dict}
\title{Dict methods}
\usage{
dict(...)

as.dict(x)

is.dict(x)

## S3 method for class 'Dict'
add(x, key, value)
}
\arguments{
\item{...}{initial elements put into the \code{Dict}.}

\item{x}{\code{R} object of \code{ANY} type for \code{\link[=as.dict]{as.dict()}} and \code{\link[=is.dict]{is.dict()}}
or of class \code{Dict} for the \code{S3} methods.}
}
\description{
The \code{\link[=dict]{dict()}} resembles Python's dict type, and is implemented
as a specialized associative \code{\link[=Container]{Container()}} thus sharing all \link[=Container]{container()}
methods with some of them being overridden to account for the associative
key-value pair semantic.
}
\details{
For a full list of all dict methods see \code{\link[=Dict]{Dict()}}.

While the \code{\link[=Dict]{Dict()}} class is based on the \code{R6} framework and
provides reference semantics, the methods described here provide an \code{S3}
interface with copy semantics. Note that any \code{S3} methods defined for the
\code{Container} class also work with \code{Dict} objects, even if not overwritten
explicitly.
\subsection{Methods}{
}

\itemize{
\item \code{dict(...)} initializes and returns an object of class \code{Dict}
}

\itemize{
\item \code{as.dict(x)} coerces \code{x} to a dictionary
}

\itemize{
\item \code{is.dict(x)} returns \code{TRUE} if \code{x} is of class \code{Dict}
and \code{FALSE} otherwise.
}

\itemize{
\item \code{add(x, key, value)} adds key-value pair to \code{x}. If the \code{key}
already existed, an error is given.
}
}
\examples{
d = dict(a = 1, b = "one", f = mean)
print(d)
print(values(d))

\dontrun{
dict(a = 1, 2)                       # all elements must be named}

# Coercion
as.dict(list(A = 1:3, B = "b"))
as.dict(c(x = 1, y = "x", z = 2 + 3))
}
\seealso{
See \code{\link[=container]{container()}} for all inherited methods. For the full class
documentation see \code{\link[=Dict]{Dict()}} and it's superclass \link[=container]{Container()}.
}
