% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictS3.R
\name{dictS3}
\alias{dictS3}
\alias{dict}
\alias{as.dict}
\alias{is.dict}
\alias{keys}
\alias{names.Dict}
\title{Dictionary}
\usage{
dict(...)

as.dict(x, ...)

is.dict(x)

keys(x)

\method{names}{Dict}(x)
}
\arguments{
\item{...}{initial elements put into the \code{Dict}.}

\item{x}{any \code{R} object, or an object inheriting from class 'Dict' or
'Container'.}
}
\value{
\code{\link[=dict]{dict()}} returns a \link[=dict]{Dict()} object.

\code{\link[=as.dict]{as.dict()}} coerces to a dict.

\code{\link[=is.dict]{is.dict()}} returns \code{TRUE} if its argument is a \code{\link[=Dict]{Dict()}} and \code{FALSE}
otherwise.

\code{keys()} returns a \code{character} vector of all the dict's keys.

\code{names()} returns the same as \code{\link[=keys]{keys()}}.
}
\description{
The \code{\link[=dict]{dict()}} resembles Python's dict type, and is implemented
as a specialized associative \code{\link[=container]{container()}} thus sharing all \code{\link[=container]{container()}}
methods with some of them being overridden to account for the associative
key-value pair semantic.
}
\details{
For a full list of all dict methods see \code{\link[=Dict]{Dict()}}.
}
\examples{
dict(a = 1, b = "one", f = mean)

d1 = dict(a = 1, b = 2)
mode(values(d1))                     # 'list'
d2 = dict(c(a = 1, b = 2))
mode(values(d2))                     # 'numeric'

\dontrun{
dict(a = 1, 2)                       # all elements must be named}

# Initialize from data.frame
daf = data.frame(A = 1:3, B = 3:1)
d = dict(daf)
d

# Coerce from other types
as.dict(list(A = 1:3, B = "b"))
as.dict(c(x = 1, y = "x", z = 2 + 3))

d = dict(x = 1, y = 2)
keys(d)
names(d)
}
\seealso{
\code{\link[=Dict]{Dict()}}, \code{\link[=container]{container()}}
}
