% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict-S3.R
\name{dictS3binOp}
\alias{dictS3binOp}
\alias{+.Dict}
\alias{-.Dict}
\title{Binary dict operators}
\usage{
\method{+}{Dict}(d1, d2)

\method{-}{Dict}(d1, d2)
}
\arguments{
\item{d1}{\code{\link[container]{Dict}} object}

\item{d2}{\code{\link[container]{Dict}} object}
}
\value{
\code{\link[container]{Dict}} object
}
\description{
Binary operators for \code{Dict} objects.
}
\details{
\code{d1 + d2}: return a copy of \code{d1} updated by \code{d2}.

\code{d1 - d2}: return a copy of \code{d1} with all keys being
 removed that occured in \code{d2}.
}
