% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict-S3.R
\name{dictS3binOp}
\alias{dictS3binOp}
\alias{+.Dict}
\alias{-.Dict}
\title{Binary \code{Dict} operators}
\usage{
\method{+}{Dict}(d1, d2)

\method{-}{Dict}(d1, d2)
}
\arguments{
\item{d1}{\code{\link[=Dict]{Dict()}} object}

\item{d2}{\code{\link[=Dict]{Dict()}} object}
}
\value{
For \code{+} returns a copy of \code{d1} updated by \code{d2}.

For \code{-} returns a copy of \code{d1} with all \code{d2} keys being discarded.
}
\description{
Binary operators for \code{\link[=Dict]{Dict()}} objects.
}
