% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict-S3.R
\name{dictS3replace}
\alias{dictS3replace}
\alias{[[.Dict}
\alias{[.Dict}
\alias{[[<-.Dict}
\alias{[<-.Dict}
\title{Extract or replace parts of a \code{Dict}}
\usage{
\method{[[}{Dict}(x, key, default = NULL)

\method{[}{Dict}(x, key, default = NULL)

\method{[[}{Dict}(x, key, add = FALSE) <- value

\method{[}{Dict}(x, key, add = FALSE) <- value
}
\arguments{
\item{x}{\code{\link[=Dict]{Dict()}} object.}

\item{key}{\code{character} name of elements to extract or replace.}

\item{default}{A suitable default value.}

\item{add}{\code{logical} If \code{FALSE} and key is not yet in the dictionary, an
error is signaled. This is different from standard R lists, where instead
just a new entry would be generated. To behave like standard R lists, set
\code{add = TRUE}, which also sets the value at the key, but will also add a
new key-value pair if the key is not yet in the dictionary.}

\item{value}{A suitable replacement value.}
}
\value{
For \code{[[} returns the element found at key, or \code{default} if not found.

For \code{[} returns a dictionary containing the extracted values.

For \verb{[[<-} replaces the value associated with \code{key}. If \code{key} is not
in the dictionary, by default, an error is raised, unless \code{add} was set to
\code{TRUE}, in which case the value is added to the dictionary.

For \verb{[<-} replaces the values at the given \code{keys}.
}
\description{
Access and assignment operators for \code{\link[=Dict]{Dict()}} objects.
}
