% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discard.R
\name{discard}
\alias{discard}
\alias{ref_discard}
\alias{discard.Container}
\alias{ref_discard.Container}
\title{Discard elements}
\usage{
discard(.x, ...)

ref_discard(.x, ...)

\method{discard}{Container}(.x, ...)

\method{ref_discard}{Container}(.x, ...)
}
\arguments{
\item{.x}{any \code{R} object.}

\item{...}{elements to be discarded.}
}
\value{
For \code{Container}, an object of class \code{Container} (or one of the
respective derived classes).
}
\description{
Search and remove an element from an object. If the element is not found,
ignore the attempt.
}
\examples{

s = setnew("a", num = 1:3, data = iris)
print(s)
discard(s, 1:3, "a")
discard(s, iris)
discard(s, "b")  # ignored
}
