% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discard.R
\name{discard}
\alias{discard}
\alias{discard_}
\alias{discard.Container}
\alias{discard_.Container}
\alias{discard.dict.table}
\title{Discard elements}
\usage{
discard(x, ...)

discard_(x, ...)

\method{discard}{Container}(x, ...)

\method{discard_}{Container}(x, ...)

\method{discard}{dict.table}(x, ...)
}
\arguments{
\item{x}{any \code{R} object.}

\item{...}{elements to be discarded. For \code{Container}, \code{Deque} and \code{Set}
objects these will be elements contained in the objects. For \code{Dict} these
are key names. For \code{dict.table} these can be either column names or column
indices or both.}

\item{column}{\code{character} name or \code{numeric} index of column.}
}
\value{
For \code{Container}, an object of class \code{Container} (or one of the
respective derived classes).

For \code{dict.table}, an object of class \code{dict.table}.
}
\description{
Search and remove an element from an object. If the element is not found,
ignore the attempt.
}
