% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ops-extract.R
\name{ExtractContainer}
\alias{ExtractContainer}
\title{Extract or Replace Parts of a Container Object}
\usage{
x[i]
x[[i]]
x[...]

x[i] <- value
x[[i]] <- value
x$name <- value
}
\arguments{
\item{x}{\code{Container} object for which to extract or replace elements.}

\item{i, ...}{indices specifying elements to extract or replace. Indices
are \code{numeric} or \code{character} vectors or a \code{list} containing both.}

\item{name}{\code{character} string (possibly backtick quoted)}

\item{value}{the replacing value of \code{ANY} type}
}
\description{
Extract or replace parts of a \code{Container} object similar
to R's base extract and replace operators on lists.
}
\details{
\code{[} selects multiple values. The indices can be \code{numeric} or
\code{character} or both. They can be passed as a \code{vector} or \code{list} or,
for convenience, just as a comma-separated sequence (see Examples).
Non-existing indices are ignored.

\code{[[} selects a single value using a \code{numeric} or \code{character} index.

\verb{[<-} replaces multiple values. The indices can be \code{numeric} or
\code{character} or both. They can be passed as a \code{vector} or \code{list}. Values can
be added by 'replacing' at new indices, which only works for \code{character}
indices.

\verb{[[<-} replaces a single value at a given \code{numeric} or \code{character} index.
Instead of an index, it is also possible to pass an element as a \code{list}. In
this case, the object is searched for the element and then the element is
replaced by the value (see Examples).

\verb{$<-} replaces a single element at a given name.
}
\examples{
co = container(a = 1, b = 2, c = 3, d = 4)
co[1:2]
co[1, 4]
co["d", 2]
co[list("d", 2)]
co[0:10]

co = container(a = 1, b = 2)
co[[1]]
co[["a"]]
co[["x"]]

co = container(a = 1, b = 2)
co[[1]] <- 9
co[["b"]] <- 8
co[["x"]] <- 7
co$z <- 99
print(co)

# Replace 8 by 0
co[[{8}]] <- 0
print(co)
}
