% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getval.R
\name{getval}
\alias{getval}
\alias{getval.Dict}
\alias{getval.dict.table}
\title{Strict element access}
\usage{
getval(x, ...)

\method{getval}{Dict}(x, key)

\method{getval}{dict.table}(x, key)
}
\arguments{
\item{x}{any \code{R} object.}

\item{...}{additional arguments to be passed to or from methods.}

\item{key}{one or more indices to be accessed. For \code{Dict} objects the keys
must be of type \code{character} while they can be \code{character} or \code{numeric} for
\code{dict.table} objects.}
}
\value{
If the \code{key} is of length one, the raw value associated with the
\code{key} is returned, otherwise an object of the same type as the one that was
accessed.
}
\description{
Try to access element and signal an error if not found.
}
\examples{

d = dict(a = 1, b = 1:3)
getval(d, "b")
getval(d, c("a", "b"))
getvalues(d, "a", "b")

dit = dict.table(a = 1:3, b = 4:6)
getval(dit, "b")
getval(dit, c("a", "b")) # or
getvalues(dit, "a", "b")
}
