% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getval.R
\name{getval}
\alias{getval}
\alias{getvalues}
\alias{getval.Dict}
\alias{getvalues.Dict}
\alias{getval.dict.table}
\alias{getvalues.dict.table}
\title{Strict element access}
\usage{
getval(.x, ...)

getvalues(.x, ...)

\method{getval}{Dict}(.x, key)

\method{getvalues}{Dict}(.x, ...)

\method{getval}{dict.table}(.x, key)

\method{getvalues}{dict.table}(.x, ...)
}
\arguments{
\item{.x}{any \code{R} object.}

\item{...}{keys or indices to be accessed or additional arguments to be
passed to or from methods.}

\item{key}{one or more indices to be accessed. For \code{Dict} objects, keys
are of type \code{character} while they can be either \code{character} or \code{numeric} for
\code{dict.table} objects.}
}
\value{
If the \code{key} is of length one, the raw value associated with the
\code{key} is returned, otherwise an object of the same type as the one that was
accessed.
}
\description{
Try to access elements and signal errors for missing elements.
}
\examples{

d = dict(a = 1, b = 1:3)
getval(d, "b")
getval(dit, c("a", "b")) # or alternatively:
getvalues(d, "a", "b")

dit = dict.table(a = 1:3, b = 4:6)
getval(dit, "b")
getval(dit, c("a", "b")) # or alternatively
getvalues(dit, "a", "b")
}
\seealso{
\code{\link[=peek]{peek()}} for less strict element access.
}
