% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getval.R
\name{getval}
\alias{getval}
\alias{getval.Dict}
\title{Strict element access}
\usage{
getval(x, ...)

\method{getval}{Dict}(x, key)
}
\arguments{
\item{x}{any \code{R} object.}

\item{...}{additional arguments to be passed to or from methods.}

\item{key}{\code{character} name of key.}
}
\value{
For \code{Dict} returns value at \code{key}, if \code{key} is found in \code{Dict}
otherwise throws an error.
}
\description{
Try to access elements and signal an error if not found.
}
