% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has.R
\name{has}
\alias{has}
\alias{has.Container}
\alias{has.Dict}
\title{Determine if object has some element}
\usage{
has(x, ...)

\method{has}{Container}(x, elem)

\method{has}{Dict}(x, key)
}
\arguments{
\item{x}{any \code{R} object.}

\item{...}{additional arguments to be passed to or from methods.}

\item{elem}{element to search for.}

\item{key}{\code{character} name of key to search for.}
}
\value{
For \code{Container} \code{TRUE} if element is in container otherwise \code{FALSE}.

For \code{Dict} \code{TRUE} if key is in dict otherwise \code{FALSE}.
}
\description{
Determine if object has some element
}
