% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has.R
\name{has}
\alias{has}
\alias{has.Dict}
\alias{has.dict.table}
\title{Determine if object has some element}
\usage{
has(x, ...)

\method{has}{Dict}(x, key)

\method{has}{dict.table}(x, column)
}
\arguments{
\item{x}{any \code{R} object.}

\item{...}{additional arguments to be passed to or from methods.}

\item{key}{\code{character} name of key to search for.}

\item{column}{\code{character} name or \code{numeric} index of column.}
}
\value{
For \code{Dict} \code{TRUE} if key is in dict otherwise \code{FALSE}.

For \code{dict.table} \code{TRUE} if column name or index is in dict.table,
otherwise \code{FALSE}.
}
\description{
Determine if object has some element
}
