% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IteratorR6.R
\name{is.subsettable}
\alias{is.subsettable}
\title{Check if object is subsettable}
\usage{
is.subsettable(x, .subset = .subset2)
}
\arguments{
\item{x}{any \code{R} object}

\item{.subset}{subset function to be used on \code{x}}
}
\value{
returns \code{TRUE} if object is subsettable otherwise \code{FALSE}
}
\description{
Check if object is subsettable
}
