% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterator.R
\docType{class}
\name{Iterable}
\alias{Iterable}
\title{Iterable abstract class interface}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Iterable
}
\description{
An \code{Iterable} is an object that provides an \code{iter()}
method, which is expected to return an \code{Iterator} object. This class
defines the abstract class interface such that each class inheriting this
class provides an \code{iter()} method and must implement a private method
\code{create_iter}, which must return an \code{\link[container]{Iterator}} object.
}
\section{Inherited methods}{

Inherits method \code{iter} from abstract \code{\link[container]{Iterable}} class.
}

\section{Iterable method/interface}{

\describe{
 \item{\code{iter()}}{Return \code{\link[container]{Iterator}} object.}
}
}

\seealso{
\code{\link[container]{Iterator}} and \code{\link[container]{Container}}
}
\author{
Roman Pahl
}
\keyword{datasets}
