% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterator.R
\docType{class}
\name{Iterator}
\alias{Iterator}
\title{Iterator}
\description{
An \code{Iterator} is an object that allows to iterate over
sequences. It implements \code{.next} and \code{get} to iterate and retrieve the
value of the sequence it is associated with.
}
\author{
Roman Pahl
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Iterator$new()}}
\item \href{#method-begin}{\code{Iterator$begin()}}
\item \href{#method-get}{\code{Iterator$get()}}
\item \href{#method-get_next}{\code{Iterator$get_next()}}
\item \href{#method-pos}{\code{Iterator$pos()}}
\item \href{#method-has_next}{\code{Iterator$has_next()}}
\item \href{#method-.next}{\code{Iterator$.next()}}
\item \href{#method-print}{\code{Iterator$print()}}
\item \href{#method-clone}{\code{Iterator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\code{Iterator} constructor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Iterator$new(x = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{sequence to iterate over}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisibly returns \code{Iterator} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-begin"></a>}}
\if{latex}{\out{\hypertarget{method-begin}{}}}
\subsection{Method \code{begin()}}{
set iterator to start of sequence
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Iterator$begin()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get"></a>}}
\if{latex}{\out{\hypertarget{method-get}{}}}
\subsection{Method \code{get()}}{
get element where the iterator points to
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Iterator$get()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_next"></a>}}
\if{latex}{\out{\hypertarget{method-get_next}{}}}
\subsection{Method \code{get_next()}}{
get next element
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Iterator$get_next()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pos"></a>}}
\if{latex}{\out{\hypertarget{method-pos}{}}}
\subsection{Method \code{pos()}}{
get iterator position
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Iterator$pos()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-has_next"></a>}}
\if{latex}{\out{\hypertarget{method-has_next}{}}}
\subsection{Method \code{has_next()}}{
check if \code{iterator} has more elements
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Iterator$has_next()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{TRUE} if \code{iterator} has next element else \code{FALSE}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-.next"></a>}}
\if{latex}{\out{\hypertarget{method-.next}{}}}
\subsection{Method \code{.next()}}{
increment \code{iterator}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Iterator$.next()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
print method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Iterator$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Iterator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
