% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterator-S3.R
\name{iterS3}
\alias{iterS3}
\alias{iter}
\alias{iter.default}
\alias{is.iterator}
\alias{itbegin}
\alias{itget}
\alias{itget_next}
\alias{itpos}
\alias{ithas_next}
\alias{itnext}
\title{Iterator S3 methods}
\usage{
iter(x)

\method{iter}{default}(x)

is.iterator(x)

itbegin(it)

itget(it)

itget_next(it)

itpos(it)

ithas_next(it)

itnext(it)
}
\arguments{
\item{it}{\code{Iterator} object}
}
\description{
An \code{Iterator} is an object that allows to iterate over
sequences. It implements \code{.next} and \code{get} to iterate and retrieve the
value of the sequence it is associated with.
}
\section{S3 methods for class \code{Iterator}}{

\describe{
\item{\code{itbegin(it)}}{Reset iterator position to 1.}
\item{\code{itget(it)}}{Get value at current iterator position.}
\item{\code{itget_next()}}{Get value after incrementing by one.}
\item{\code{itpos()}}{Return current iterator position.}
\item{\code{ithas_next(it)}}{Return TRUE if there is a next element.}
\item{\code{itnext(it)}}{Increment iterator to point at next element.}
}
}

\examples{
it <- iter(list("A", 1, 2))
while(ithas_next(it)) {
  print(itget_next(it))
}
ithas_next(it)   # FALSE
print(it)       # <Iterator> at position 3
itbegin(it)
print(it)       # <Iterator> at position 0
}
