% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-IteratorS3.R
\name{iterS3}
\alias{iterS3}
\alias{iter}
\alias{iter.Container}
\alias{iter.default}
\alias{is.iterator}
\alias{is.iterable}
\alias{begin}
\alias{get_value}
\alias{get_next}
\alias{has_next}
\alias{has_value}
\alias{pos}
\alias{next_iter}
\alias{reset_iter}
\alias{length.Iterator}
\title{Iterate over Sequences}
\usage{
iter(x, ...)

\method{iter}{Container}(x, ...)

\method{iter}{default}(x, ...)

is.iterator(x)

is.iterable(x)

begin(it)

get_value(it)

get_next(it)

has_next(it)

has_value(it)

pos(it)

next_iter(it)

reset_iter(it)

\method{length}{Iterator}(x)
}
\arguments{
\item{x}{an object of class \code{\link[=Iterable]{Iterable()}} or any other \code{R} object. In the
latter case, \code{x} will always be coerced to a base \code{R} list prior to creating
the \code{iterator}.}

\item{...}{other parameters passed to or from methods}

\item{it}{\code{Iterator} object}
}
\value{
\code{length} returns the number of elements that can be iterated over.
}
\description{
An \code{Iterator} is an object that allows to iterate over
sequences. It implements \code{next_iter} and \code{get_value} to iterate and retrieve the
value of the sequence it is associated with.
}
\details{
For more details on the methods see \code{\link[=Iterator]{Iterator()}}.
}
\seealso{
\code{\link[=Iterator]{Iterator()}}
}
