% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iteratorS3.R
\name{iterS3}
\alias{iterS3}
\alias{iter}
\alias{iter.Container}
\alias{iter.default}
\alias{is.iterator}
\alias{begin}
\alias{get_value}
\alias{get_next}
\alias{has_next}
\alias{has_value}
\alias{pos}
\alias{next_iter}
\alias{reset_iter}
\alias{length.Iterator}
\title{Iterator}
\usage{
iter(x)

\method{iter}{Container}(x)

\method{iter}{default}(x)

is.iterator(x)

begin(it)

get_value(it)

get_next(it)

has_next(it)

has_value(it)

pos(it)

next_iter(it)

reset_iter(it)

\method{length}{Iterator}(x)
}
\arguments{
\item{x}{an iterable object}

\item{it}{\code{Iterator} object}
}
\value{
\code{length()} returns the number of elements to iterate
}
\description{
An \code{Iterator} is an object that allows to iterate over
sequences. It implements \code{next_iter} and \code{get_value} to iterate and retrieve the
value of the sequence it is associated with.
}
\details{
For more details on the methods see \code{\link[=Iterator]{Iterator()}}.
}
\seealso{
\code{\link[=Iterator]{Iterator()}}
}
