% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ops-arithmetic.R
\name{OpsArith}
\alias{OpsArith}
\alias{+.Set}
\alias{/.Set}
\alias{-.Set}
\alias{+.Container}
\alias{+.Deque}
\alias{+.Dict}
\alias{-.Dict}
\alias{/.Dict}
\title{Arithmetic Operators for Containers}
\usage{
\method{+}{Set}(x, y)

\method{/}{Set}(x, y)

\method{-}{Set}(x, y)

\method{+}{Container}(x, y)

\method{+}{Deque}(x, y)

\method{+}{Dict}(x, y)

\method{-}{Dict}(x, y)

\method{/}{Dict}(x, y)
}
\arguments{
\item{x, y}{Depending on the operator at least one or both must be of
class \code{\link[=Container]{Container()}} or the respective derived class. Both must be at least
iterable.}
}
\value{
For \code{Set}, \code{+} returns the union of both sets.

For \code{Set} , \code{/} returns the intersection of both sets.

For \code{Set}, \code{-} returns the set-difference of both sets.

For \code{Container}, \code{+} returns a new \code{\link[=Container]{Container()}} object containing
all elements from \code{x} and \code{y}. Note that x + y and y + x will yield identical
results.

For \code{Deque}, \code{+} returns a new \code{\link[=Deque]{Deque()}} object with all elements
from \code{x} and \code{y}. If \code{x} is not a \code{\link[=Deque]{Deque()}}, it's elements will be added to
the left of \code{y}.

For \code{Dict}, \code{+} returns a copy of \code{x} updated by \code{y}.

For \code{Dict}, \code{-} returns a copy of \code{x} with all \code{y} keys being
discarded, that is, it keeps only those keys of \code{x} that are not in \code{y}.

For \code{Dict}, \code{/} returns a copy of \code{x} keeping only the keys that
exist in both, that is, all keys in \code{x} that do not exist in \code{y} are
removed.
}
\description{
Binary arithmetic operators for \code{\link[=Container]{Container()}} objects and
derived classes.
}
