% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ops-extract.R
\name{OpsExtract}
\alias{OpsExtract}
\title{Extract or Replace Part of a Container Object}
\arguments{
\item{x}{\code{Container} object for which to extract or replace elements.}

\item{j}{\code{numeric} or \code{character} indices to extract.}

\item{...}{indices of elements to be extracted}

\item{name}{a literal character string or a name (possibly backtick quoted).}
}
\value{
the values at the given indices.
}
\description{
Extract or replace parts of a \code{Container} object similar
to R's base extract operators on lists.
}
