% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ops-extract.R
\name{OpsExtract}
\alias{OpsExtract}
\alias{[.Container}
\alias{[[.Container}
\title{Extract Parts of a Container Object}
\usage{
\method{[}{Container}(x, ..., .default = NULL)

\method{[[}{Container}(x, i)
}
\arguments{
\item{x}{\code{Container} object from which to extract elements.}

\item{.default}{value to be returned if peeked value does not exist.}

\item{i, ...}{indices specifying elements to extract. Indices
can be \code{logical}, \code{numeric}, \code{character}, or a mix of these.
They can be passed as \code{vector}, \code{list}, or comma-separated
sequence.
Non-existing indices are ignored, but can be substituted if a
default value was provided (see \code{.default} parameter).
Supports alpha-numeric range selection via non-standard evaluation
(e.g., \code{co[a:b]} see Examples).}
}
\description{
Extract parts of a \code{Container} object similar
to R's base extract operators on lists.
}
\details{
\code{[} selects multiple values. The indices can be \code{numeric} or
\code{character} or both. They can be passed as a \code{vector} or \code{list} or,
for convenience, just as a comma-separated sequence (see Examples).
Non-existing indices are ignored.

\code{[[} selects a single value using a \code{numeric} or \code{character} index.
}
\section{Warning}{

Alpha-numeric range selection (e.g. \code{co[a:b]}) is intended for
interactive use only, because the non-standard evaluation of the
range arguments can have unanticipated consequences.
For programming it is therefore better to use the standard subsetting
indices (e.g. \code{co[1:2]} or \code{co[c("a", "b")]}).
}

\examples{
co = container(a = 1, b = 2, c = 3, d = 4)
co[1:2]
co[1, 4]
co["d", 2]
co[list("d", 2)]
co[0:10]

# Boolean selection

# Custom default values
co[1:2, 99, .default = 0]
co[1:2, "z", .default = -1]
co[1:2, "z", .default = 3:4]

# Alpha-numeric range selection
co[a:b]
co[a:b, d:c]
co[1:c]
co[d:2]
co[a:8, .default = a]

co = container(a = 1, b = 2)
co[[1]]
co[["a"]]
co[["x"]]
}
