% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ops-extract.R
\name{OpsExtract}
\alias{OpsExtract}
\title{Extract Operators for Containers}
\arguments{
\item{x}{object from which to extract element(s)}
}
\description{
Extract parts of Container (or derived classes) objects
}
