% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ops-logic.R
\name{OpsLogic}
\alias{OpsLogic}
\title{Logic Operators for Containers}
\arguments{
\item{x, y}{Depending on the operator at least one must be of class
\code{\link[=Container]{Container()}} or the respective derived class and the other at least be
coercible to the respective class.}
}
\description{
Binary logic operators for \code{\link[=Container]{Container()}} objects and
derived classes.
}
