% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ops-replace.R
\name{OpsReplace}
\alias{OpsReplace}
\alias{[[<-.Dict}
\alias{[<-.Dict}
\alias{[[<-.Set}
\alias{[[<-.dict.table}
\title{Replace Operators for Containers}
\usage{
\method{[[}{Dict}(x, key) <- value

\method{[}{Dict}(x, key) <- value

\method{[[}{Set}(x, element) <- value

\method{[[}{dict.table}(x, j) <- value
}
\arguments{
\item{x}{\code{Container} or \code{dict.table} object in which to replace elements or}

\item{key}{\code{character} name of key or column.
columns.}

\item{value}{A suitable replacement value.}

\item{element}{an element of the \code{\link[=Set]{Set()}}}

\item{j}{\code{numeric} or \code{character} column index.}
}
\value{
For \code{Dict} \verb{[[<-} replaces the value associated with \code{key}. If \code{key}
is not in the dict, by default, an error is raised, unless the \code{add}
argument was set to \code{TRUE} (see details at \code{add} description).

For \code{Dict}, \verb{[<-} replaces the values at the given \code{keys}. If one or
more keys are not found, an error is signaled, unless \code{add} was set to
\code{TRUE}.

For \code{Set}, \verb{[[<-} replaces \code{element} by \code{value}. If the \code{element}
does not exist, an error is signaled, unless you pass \code{NULL}, in which case
the value is just added to the \code{Set}.

For \code{dict.table}, \code{[[} replaces the selected column. If the column
does not exist, an error is signaled.
}
\description{
Replace parts of \code{Container} or \code{dict.table} objects.
}
