% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ops-replace.R
\name{OpsReplace}
\alias{OpsReplace}
\alias{[[<-.Dict}
\alias{[<-.Dict}
\alias{[[<-.Set}
\title{Replace Operators for Containers}
\usage{
\method{[[}{Dict}(x, key, add = FALSE) <- value

\method{[}{Dict}(x, key, add = FALSE) <- value

\method{[[}{Set}(x, element) <- value
}
\arguments{
\item{x}{\code{Container} object in which to replace element(s)}

\item{add}{\code{logical} If \code{FALSE} and key is not yet in the dict, an
error is signaled. This is different from standard R lists, where instead
just a new entry would be generated. To behave like standard R lists, set
\code{add = TRUE}, which also sets the value at the key, but will also add a
new key-value pair if the key is not yet in the dictionary.}

\item{value}{A suitable replacement value.}

\item{element}{an element of the \code{\link[=Set]{Set()}}}
}
\value{
For \code{Dict} \verb{[[<-} replaces the value associated with \code{key}. If \code{key}
is not in the dict, by default, an error is raised, unless the \code{add}
argument was set to \code{TRUE} (see details at \code{add} description).

For \code{Dict}, \verb{[<-} replaces the values at the given \code{keys}. If one or
more keys are not found, an error is signaled, unless \code{add} was set to
\code{TRUE}.

For \code{Set}, \verb{[[<-} replaces \code{element} by \code{value}.
}
\description{
Replace parts of Container objects.
}
