% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peek.R
\name{peek}
\alias{peek}
\alias{peek.Deque}
\alias{peek.Dict}
\title{Peek element of an object}
\usage{
peek(x, ...)

\method{peek}{Deque}(x)

\method{peek}{Dict}(x, key, default = NULL)
}
\arguments{
\item{x}{any \code{R} object.}

\item{...}{additional arguments to be passed to or from methods.}

\item{key}{\code{character} key of the value to peek.}

\item{default}{\code{ANY} default value to return if \code{key} not in the dictionary.}
}
\value{
For \code{Deque} the last element at the right side of the deque.

For \code{Dict} if \code{key} does exists, the associated value is returned
otherwise the given \code{default} value.
}
\description{
Search an element in an object and return a copy of it. If there is no
element, return \code{NULL} or some default value.
}
