% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peek.R
\name{peek}
\alias{peek}
\alias{peek.Deque}
\alias{peek.Dict}
\alias{peek.dict.table}
\title{Peek element of an object}
\usage{
peek(x, ...)

\method{peek}{Deque}(x)

\method{peek}{Dict}(x, key, default = NULL)

\method{peek}{dict.table}(x, column, default = NULL)
}
\arguments{
\item{x}{any \code{R} object.}

\item{...}{additional arguments to be passed to or from methods.}

\item{key}{\code{character} key of the value to peek.}

\item{default}{\code{ANY} default value to return if \code{key} not in the dictionary.}

\item{column}{\code{character} name or \code{numeric} index of column.}
}
\value{
For \code{Deque} the last element at the right side of the deque.

For \code{Dict} if \code{key} does exists, the associated value is returned
otherwise the given \code{default} value.

For \code{dict.table} returns the column if it does exist otherwise
the given \code{default} value. If the default length does not match the number
of rows, it is recycled accordingly and a warning is given, unless the
default value had a length of 1.
}
\description{
Search an element in an object and return a copy of it. If there is no
element, return \code{NULL} or some default value.
}
