% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peek.R
\name{peek}
\alias{peek}
\alias{peekleft}
\alias{peek.Deque}
\alias{peekleft.Deque}
\alias{peek.Dict}
\alias{peek.dict.table}
\title{Peek at element}
\usage{
peek(x, ...)

peekleft(x, ...)

\method{peek}{Deque}(x, default = NULL)

\method{peekleft}{Deque}(x, default = NULL)

\method{peek}{Dict}(x, key, default = NULL)

\method{peek}{dict.table}(x, key, default = NULL)
}
\arguments{
\item{x}{an \code{R} object of the respective class.}

\item{...}{additional arguments to be passed to or from methods.}

\item{default}{value to be returned if peeked value does not exist.}

\item{key}{\code{character} name of the value to peek.}
}
\value{
For \code{Deque} the first (\code{peekleft}) or last (\code{peek}) element.

For \code{Dict}, returns the associated value if \code{key} does exists,
otherwise the given \code{default} value.

For \code{dict.table}, returns the column named \code{key} if it exist otherwise
the given \code{default} value. If the default length does not match the number
of rows, it is recycled accordingly and a warning is given, unless the
default value has a length of 1, in which case recycling is done silently.
}
\description{
Try to access element and return some default value if not found.
The \code{peek} function basically enables safe element access in cases where
elements don't exist or the accessed objects is empty, thereby freeing the
user from additional effort (like if-statements) that would be required to
catch these cases manually.
}
\details{
\code{peek} tries to access specific values.

\code{peekleft} peeks at first element of a \code{Deque}.
}
\examples{
# Deque
d = deque(1, 2, 3)
peek(d)
peekleft(d)
peek(deque())
peek(deque(), default = 0)
peekleft(deque(), default = 0)

# Dict
d = dict(a = 1, b = 1:3)
peek(d, "b")
peek(d, "x")
peek(d, "x", default = 4:7)

# dict.table
dit = dict.table(a = 1:3, b = 4:6)
peek(dit, "a")
peek(dit, 1)
peek(dit, 3)
peek(dit, "x")
peek(dit, "x", default = 0)
}
\seealso{
\code{\link[=pop]{pop()}}
}
