% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peekitem.R
\name{peekitem}
\alias{peekitem}
\alias{peekitem.Container}
\alias{peekitem.dict.table}
\title{Peek at random item}
\usage{
peekitem(x, ...)

\method{peekitem}{Container}(x, default = NULL)

\method{peekitem}{dict.table}(x, default = NULL)
}
\arguments{
\item{x}{an \code{R} object of the respective class.}

\item{...}{additional arguments to be passed to or from methods.}

\item{default}{value to be returned if collection is empty.}
}
\value{
The value that was randomly chosen from the collection of values.

For \code{dict.table}, returns a randomly chosen column. If there are no
columns, the \code{default} value is returned.
}
\description{
Try to access an element that is chosen randomly from a collection and if
not found (i.e. collection is empty), return some default value.
The \code{peekitem} function therefore can be used to sample randomly (with
replacement) from a collection of elements.
}
\examples{
co = container(1, 2, 3)
peekitem(co)
peekitem(container(), default = 9)

# dict.table
dit = dict.table(a = 1:5, b = 5:1)
peekitem(dit)
}
\seealso{
\code{\link[=peek]{peek()}}, \code{\link[=popitem]{popitem()}}
}
