% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop.R
\name{pop}
\alias{pop}
\alias{pop.Deque}
\alias{pop.Dict}
\alias{pop.dict.table}
\title{Pop element}
\usage{
pop(x, ...)

\method{pop}{Deque}(x)

\method{pop}{Dict}(x, key)

\method{pop}{dict.table}(x, column)
}
\arguments{
\item{x}{any \code{R} object.}

\item{...}{additional arguments to be passed to or from methods.}

\item{key}{\code{character} key of the value to pop. If \code{key} does exists,
the associated key-value pair is deleted and it's value returned.}

\item{column}{\code{character} name or \code{numeric} index of column.}
}
\value{
For \code{Deque} the element from the right side of the deque after it
was removed.

For \code{Dict} the value associated with the key after the key-value
pair was removed from the dict.

For \code{dict.table} the column after it was removed from the dict.table.
}
\description{
Search and remove an element from an object and return a copy of the
element. If the element is not found, signal an error.
}
