% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop.R
\name{pop}
\alias{pop}
\alias{ref_pop}
\alias{ref_popleft}
\alias{ref_pop.Deque}
\alias{ref_popleft.Deque}
\alias{ref_pop.Dict}
\alias{ref_pop.dict.table}
\title{Pop element}
\usage{
ref_pop(.x, ...)

ref_popleft(.x, ...)

\method{ref_pop}{Deque}(.x)

\method{ref_popleft}{Deque}(.x)

\method{ref_pop}{Dict}(.x, key)

\method{ref_pop}{dict.table}(.x, key)
}
\arguments{
\item{.x}{any \code{R} object.}

\item{...}{additional arguments to be passed to or from methods.}

\item{key}{\code{character} name of the value to pop.
the associated key-value pair is deleted and it's value returned.}
}
\value{
For \code{Deque} the first (\code{ref_popleft}) or last (\code{ref_pop}) element of
the deque after it was removed.

For \code{Dict} the value associated with the key after the key-value
pair was removed from the dict. If \code{key} is not found, an error is raised.

For \code{dict.table}, returns the column named \code{key} after it was
removed from the dict.table. If column does not exist, an error is raised.
}
\description{
Search and return an element and remove it afterwards from the object.
If the element is not found, signal an error.
}
\details{
All functions work by reference, that is, the original object is altered.
\code{ref_pop(.x)} tries to access specific values.

\code{ref_poplefref_popleft(.x)} pops first element of a \code{Deque}.
}
\examples{
# Deque
d = deque(1, 2, 3)
ref_pop(d)
ref_popleft(d)

\dontrun{
ref_pop(deque())  # pop at empty Deque}

# Dict
d = dict(a = 1, b = 1:3)
ref_pop(d, "b")
print(d)

\dontrun{
ref_pop(d, "x")  # key 'x' not in Dict}

# dict.table
dit = dict.table(a = 1:3, b = 4:6)
ref_pop(dit, "a")
print(dit)
\dontrun{
ref_pop(dit, "x")  # Column 'x' not in dict.table}
}
\seealso{
\code{\link[=peek]{peek()}}
}
