% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop.R
\name{popitem}
\alias{popitem}
\alias{popitem.Container}
\title{Pop random element}
\usage{
popitem(x)

\method{popitem}{Container}(x)
}
\arguments{
\item{x}{any \code{R} object.}

\item{...}{additional arguments to be passed to or from methods.}
}
\value{
For \code{Container} a randomly popped element is returned.
}
\description{
Randomly select and remove an element from an object and return a copy of the
element. If there is no element at all, signal an error.
This function can be used to destructively iterate over a collection of
elements as often used in set algorithms.
}
