% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popitem.R
\name{popitem}
\alias{popitem}
\alias{popitem_}
\alias{popitem_.Container}
\alias{popitem_.dict.table}
\title{Pop a random item}
\usage{
popitem_(.x, ...)

\method{popitem_}{Container}(.x)

\method{popitem_}{dict.table}(.x)
}
\arguments{
\item{.x}{an \code{R} object of the respective class.}

\item{...}{additional arguments to be passed to or from methods.}
}
\value{
The value that was randomly chosen from the collection of values.
}
\description{
Randomy select and return an element and remove it afterwards. If there is
no such element, signal an error.
The \code{popitem} function can be used to sample randomly (without replacement)
from a collection of elements and this way to destructively iterate over
collections as often used in set algorithms.
}
\examples{
co = container(1, 2, 3)
while (!is_empty(co))
    print(popitem_(co))

# dict.table
dit = dict.table(a = 1:3, b = 3:1)
while (!is_empty(dit))
    print(popitem_(dit))
}
\seealso{
\code{\link[=pop]{pop()}}, \code{\link[=peekitem]{peekitem()}}
}
