% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace.R
\name{replace}
\alias{replace}
\alias{replace_}
\alias{replace.Container}
\alias{replace_.Container}
\alias{replace.Dict}
\alias{replace_.Dict}
\title{Strict element replacement}
\usage{
replace(.x, ...)

replace_(.x, ...)

\method{replace}{Container}(.x, old, new, add = FALSE)

\method{replace_}{Container}(.x, old, new, add = FALSE)

\method{replace}{Dict}(.x, old, new)

\method{replace_}{Dict}(.x, old, new)
}
\arguments{
\item{.x}{any \code{R} object.}

\item{...}{additional arguments to be passed to or from methods.}

\item{old}{old element to be found and replaced.}

\item{new}{the new element replacing the old one.}

\item{add}{\code{logical} if \code{FALSE} (default) and element was not found,
an error is given. In contrast, if set to \code{TRUE} the new element is added
regardless of whether it is used as a replacement for an existing element or
just added as a new element.}
}
\value{
For \code{Container}, an object of class \code{Container} (or one of the
respective derived classes).

For \code{Dict} an object of class \code{Dict}.
}
\description{
Try to find and replace elements and signal an error if not
found, unless it is stated to explicitly add the element (see option \code{add}).
}
\details{
\code{replace} uses copy semantics while \code{replace_} works by reference.
}
\examples{

co = container("x", 9)
replace(co, 9, 0)
replace(co, "x", 0)
\dontrun{
replace(co, "z", 0)              # old element ("z") is not in Container}
replace(co, "z", 0, add = TRUE)  # just add the zero without replacement


d = dict(a = 1)
replace(d, "a", 1:5)
\dontrun{
replace(d, "b", 2)              # key 'b' not in Dict}
replace(d, "b", 2, add = TRUE)  # ok, adds value
}
