% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetR6.R
\name{Set}
\alias{Set}
\title{Set}
\description{
The \code{\link[=Set]{Set()}} is considered and implemented as a specialized
\code{\link[=Container]{Container()}}, that is, elements are always unique in the \code{\link[=Container]{Container()}} and
it provides typical set operations such as \code{union} and \code{intersect}.
}
\seealso{
\code{\link[=Container]{Container()}}, \code{\link[=set]{set()}}
}
\section{Super classes}{
\code{\link[container:Iterable]{container::Iterable}} -> \code{\link[container:Container]{container::Container}} -> \code{Set}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Set$new()}}
\item \href{#method-add}{\code{Set$add()}}
\item \href{#method-peek}{\code{Set$peek()}}
\item \href{#method-replace}{\code{Set$replace()}}
\item \href{#method-diff}{\code{Set$diff()}}
\item \href{#method-intersect}{\code{Set$intersect()}}
\item \href{#method-union}{\code{Set$union()}}
\item \href{#method-is_equal}{\code{Set$is_equal()}}
\item \href{#method-is_subset}{\code{Set$is_subset()}}
\item \href{#method-is_proper_subset}{\code{Set$is_proper_subset()}}
\item \href{#method-values}{\code{Set$values()}}
\item \href{#method-clone}{\code{Set$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="container" data-topic="Iterable" data-id="iter">}\href{../../container/html/Iterable.html#method-iter}{\code{container::Iterable$iter()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="container" data-topic="Container" data-id="clear">}\href{../../container/html/Container.html#method-clear}{\code{container::Container$clear()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="container" data-topic="Container" data-id="count">}\href{../../container/html/Container.html#method-count}{\code{container::Container$count()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="container" data-topic="Container" data-id="delete">}\href{../../container/html/Container.html#method-delete}{\code{container::Container$delete()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="container" data-topic="Container" data-id="discard">}\href{../../container/html/Container.html#method-discard}{\code{container::Container$discard()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="container" data-topic="Container" data-id="empty">}\href{../../container/html/Container.html#method-empty}{\code{container::Container$empty()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="container" data-topic="Container" data-id="get_compare_fun">}\href{../../container/html/Container.html#method-get_compare_fun}{\code{container::Container$get_compare_fun()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="container" data-topic="Container" data-id="has">}\href{../../container/html/Container.html#method-has}{\code{container::Container$has()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="container" data-topic="Container" data-id="is_empty">}\href{../../container/html/Container.html#method-is_empty}{\code{container::Container$is_empty()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="container" data-topic="Container" data-id="length">}\href{../../container/html/Container.html#method-length}{\code{container::Container$length()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="container" data-topic="Container" data-id="print">}\href{../../container/html/Container.html#method-print}{\code{container::Container$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="container" data-topic="Container" data-id="remove">}\href{../../container/html/Container.html#method-remove}{\code{container::Container$remove()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="container" data-topic="Container" data-id="size">}\href{../../container/html/Container.html#method-size}{\code{container::Container$size()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="container" data-topic="Container" data-id="type">}\href{../../container/html/Container.html#method-type}{\code{container::Container$type()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\code{Set} constructor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Set$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{initial elements put into the \code{Set}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns the \code{Set} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add"></a>}}
\if{latex}{\out{\hypertarget{method-add}{}}}
\subsection{Method \code{add()}}{
Add element
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Set$add(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{elements to be added to the \code{Set}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the \code{Set} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-peek"></a>}}
\if{latex}{\out{\hypertarget{method-peek}{}}}
\subsection{Method \code{peek()}}{
peek last item
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Set$peek(default = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{default}}{returned default value if \code{Set} is empty.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
last element (according to internal order) in the \code{Set}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-replace"></a>}}
\if{latex}{\out{\hypertarget{method-replace}{}}}
\subsection{Method \code{replace()}}{
Search for occurence of \code{elem} in the \code{Set} and
replace it by \code{new}. If \code{elem} does not exist, an error is
signaled, unless \code{add} was set to \code{TRUE}, in which case \code{new} is
added.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Set$replace(old, new, add = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{old}}{element to be replaced}

\item{\code{new}}{element to be put instead of old}

\item{\code{add}}{\code{logical} if \code{TRUE} the \code{new} element is added in case
\code{old} does not exists.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the \code{Set} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-diff"></a>}}
\if{latex}{\out{\hypertarget{method-diff}{}}}
\subsection{Method \code{diff()}}{
\code{Set} difference
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Set$diff(s)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{s}}{\code{Set} object to 'subtract'}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the \code{Set} object updated as a result of the set difference
between this and s.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-intersect"></a>}}
\if{latex}{\out{\hypertarget{method-intersect}{}}}
\subsection{Method \code{intersect()}}{
\code{Set} intersection
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Set$intersect(s)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{s}}{\code{Set} object to 'intersect'}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the \code{Set} object as a result of the intersection of this and s.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-union"></a>}}
\if{latex}{\out{\hypertarget{method-union}{}}}
\subsection{Method \code{union()}}{
\code{Set} union
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Set$union(s)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{s}}{\code{Set} object to be 'unified'}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the \code{Set} object as a result of the union of this and s.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-is_equal"></a>}}
\if{latex}{\out{\hypertarget{method-is_equal}{}}}
\subsection{Method \code{is_equal()}}{
\code{Set} equality
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Set$is_equal(s)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{s}}{\code{Set} object to compare against}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{TRUE} if this is equal to \code{s}, otherwise \code{FALSE}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-is_subset"></a>}}
\if{latex}{\out{\hypertarget{method-is_subset}{}}}
\subsection{Method \code{is_subset()}}{
\code{Set} proper subset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Set$is_subset(s)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{s}}{\code{Set} object to compare against}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{TRUE} if this is subset of \code{s}, otherwise \code{FALSE}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-is_proper_subset"></a>}}
\if{latex}{\out{\hypertarget{method-is_proper_subset}{}}}
\subsection{Method \code{is_proper_subset()}}{
\code{Set} subset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Set$is_proper_subset(s)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{s}}{\code{Set} object to compare against}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{TRUE} if this is proper subset of \code{s}, otherwise \code{FALSE}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-values"></a>}}
\if{latex}{\out{\hypertarget{method-values}{}}}
\subsection{Method \code{values()}}{
Get \code{Set} values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Set$values()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
elements of the set as a base list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Set$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
