% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-S3.R
\name{setS3}
\alias{setS3}
\alias{setnew}
\alias{as.set}
\alias{is.set}
\alias{add.Set}
\title{Set S3 interface}
\usage{
setnew(...)

as.set(x)

is.set(x)

\method{add}{Set}(x, elem, ...)
}
\arguments{
\item{...}{initial elements to initialize the \code{\link[=Set]{Set()}} object or further
arguments depending on the method.}

\item{x}{\code{\link[=Set]{Set()}} object}

\item{elem}{some element of the set}
}
\description{
The \code{\link[=Set]{Set()}} is considered and implemented as a specialized
\code{\link[=Container]{Container()}}, that is, elements are always unique in the \code{\link[=Container]{Container()}} and
it provides typical set operations such as \code{union} and \code{intersect}.
}
\details{
For a detailed documentation of all methods see \code{\link[=Set]{Set()}}.
}
\seealso{
\code{\link[=Container]{Container()}}, \code{\link[=+.Set]{+.Set()}}, \code{\link[=/.Set]{/.Set()}}, \code{\link[=-.Set]{-.Set()}}, \code{\link[=<.Set]{<.Set()}},
\code{\link[=>.Set]{>.Set()}}
}
