% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-setS3.R
\name{setS3}
\alias{setS3}
\alias{setnew}
\alias{as.set}
\alias{is.set}
\title{Set}
\usage{
setnew(...)

as.set(x)

is.set(x)
}
\arguments{
\item{...}{initial elements put into the \code{Set}.}

\item{x}{\code{R} object of \code{ANY} type for \code{\link[=as.set]{as.set()}} and \code{\link[=is.set]{is.set()}}
or of class \code{Set} for the \code{S3} methods.}
}
\description{
The \code{\link[=Set]{Set()}} is considered and implemented as a specialized
\code{\link[=Container]{Container()}}, that is, \code{Set} elements are always unique. It provides
typical set operations such as \code{union} and \code{intersect}.
}
\details{
\code{Set} S3 methods:
\itemize{
\item \code{setnew(...)} initializes and returns a \code{\link[=Set]{Set()}} object.
}

\itemize{
\item \code{as.set(x)} coerces \code{x} to a set.
}

\itemize{
\item \code{is.set(x)} returns \code{TRUE} if \code{x} is of class \code{Set} and \code{FALSE}
otherwise.
}
}
\seealso{
\code{\link[=Set]{Set()}}, \code{\link[=container]{container()}}
}
