% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-setS3.R
\name{setS3}
\alias{setS3}
\alias{setnew}
\alias{as.set}
\alias{is.set}
\title{Set}
\usage{
setnew(...)

as.set(x, ...)

is.set(x)
}
\arguments{
\item{...}{initial elements put into the \code{Set}.}

\item{x}{any \code{R} object, or an object inheriting from class 'Set' or
'Container' for the corresponding S3 methods.}
}
\value{
\code{\link[=setnew]{setnew()}} returns a \code{\link[=Set]{Set()}} object.

\code{\link[=as.set]{as.set()}} coerces to a set.

\code{\link[=is.set]{is.set()}} returns \code{TRUE} if its argument is a \code{\link[=Set]{Set()}} and \code{FALSE}
otherwise.
}
\description{
The \code{\link[=Set]{Set()}} is considered and implemented as a specialized
\code{\link[=Container]{Container()}}, that is, elements are always unique in the \code{\link[=Container]{Container()}} and
it provides typical set operations such as \code{union} and \code{intersect}.
}
\details{
For a full list of all set methods see \code{\link[=Set]{Set()}}.
}
\seealso{
\code{\link[=Set]{Set()}}, \code{\link[=container]{container()}}
}
