% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-S3.R
\name{setS3}
\alias{setS3}
\alias{setnew}
\alias{as.set}
\alias{as.set.Container}
\alias{is.set}
\alias{\%e\%}
\title{Set}
\usage{
setnew(...)

as.set(x, ...)

\method{as.set}{Container}(x)

is.set(x)

e \%e\% x
}
\arguments{
\item{...}{initial elements put into the \code{Set}.}

\item{x}{any \code{R} object, or an object inheriting from class 'Set' or
'Container' for the corresponding S3 methods.}

\item{e}{some element of any type}
}
\value{
\code{\link[=setnew]{setnew()}} returns a \code{\link[=Set]{Set()}} object.

\code{\link[=as.set]{as.set()}} coerces to a set.

\code{\link[=is.set]{is.set()}} returns \code{TRUE} if its argument is a \code{\link[=Set]{Set()}} and \code{FALSE}
otherwise.

\verb{\%e\%} returns \code{TRUE} if e is an element of \code{x}
}
\description{
The \code{\link[=Set]{Set()}} is considered and implemented as a specialized
\code{\link[=Container]{Container()}}, that is, elements are always unique in the \code{\link[=Container]{Container()}} and
it provides typical set operations such as \code{union} and \code{intersect}.
}
\details{
For a full list of all set methods see \code{\link[=Set]{Set()}}.
}
\seealso{
\code{\link[=Set]{Set()}}, \code{\link[=container]{container()}}
}
