% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-S3.R
\name{setS3binOp}
\alias{setS3binOp}
\alias{+.Set}
\alias{/.Set}
\alias{-.Set}
\alias{==.Set}
\alias{<.Set}
\alias{>.Set}
\title{Binary set operators}
\usage{
\method{+}{Set}(s1, s2)

\method{/}{Set}(s1, s2)

\method{-}{Set}(s1, s2)

\method{==}{Set}(s1, s2)

\method{<}{Set}(s1, s2)

\method{>}{Set}(s1, s2)
}
\arguments{
\item{s1}{\code{\link[container]{Set}} object}

\item{s2}{\code{\link[container]{Set}} object}
}
\value{
union of both sets

intersection of both sets

set-difference of both sets

\code{TRUE} if both sets are equal, else \code{FALSE}

\code{TRUE} if s1 is subset of s2, else \code{FALSE}

\code{TRUE} if s1 is superset of s2, else \code{FALSE}
}
\description{
Binary operators for \code{Set} objects.
}
