% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpack.R
\name{unpack}
\alias{unpack}
\title{Unpack nested objects}
\usage{
unpack(x)
}
\arguments{
\item{x}{any \code{R} object}
}
\value{
a \code{list}
}
\description{
Similary to \code{\link[=unlist]{unlist()}} recursively unpacks any (possibly nested) structure
into a flat list. In contrast to \code{\link[=unlist]{unlist()}}, \code{\link[=unpack]{unpack()}} also works with
(possibly nested) \code{\link[=Container]{Container()}} objects.
}
