% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update}
\alias{update}
\alias{update.Dict}
\alias{update.list}
\title{Update elements}
\usage{
update(x, other, ...)

\method{update}{Dict}(x, other)

\method{update}{list}(x, other)
}
\arguments{
\item{x}{any \code{R} object}

\item{other}{any object of the same type as \code{x}}

\item{...}{additional arguments to be passed to or from methods.}
}
\value{
For \code{Dict} adds elements of \code{other} dict for keys not yet in
the dict and replaces the values of existing keys otherwise.

For \code{list} adds elements of \code{other} for names not yet in the
list and replaces the values of existing names otherwise.
}
\description{
Takes an object and updates it with values from another object
by replacing the values at existing keys and adding values for new keys of
the other object. For example, this can be used to update some parameter
lists.
}
